% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes_from_df.R,
%   R/attributes_from_spreadsheet.R
\name{attributes_from_df}
\alias{attributes_from_df}
\alias{attributes_from_spreadsheet}
\title{Produce ROCK attributes in YAML format from a data frame or spreadsheet}
\usage{
attributes_from_df(
  x,
  classInstanceId_col,
  colsToInclude = names(x),
  output = NULL,
  delimiterString = rock::opts$get("delimiterString"),
  attributeContainer = rock::opts$get("attributeContainers")[1]
)

attributes_from_spreadsheet(
  x,
  classInstanceId_col,
  colsToInclude = names(x),
  col_renaming = NULL,
  sheet = 1,
  output = NULL,
  delimiterString = rock::opts$get("delimiterString"),
  attributeContainer = rock::opts$get("attributeContainers")[1]
)
}
\arguments{
\item{x}{For \code{attributes_from_df}, the data frame; and
for \code{attributes_from_spreadsheet}, the path to the spreadsheet
(an argument to \code{\link[=read_spreadsheet]{read_spreadsheet()}}).}

\item{classInstanceId_col}{The column name of the column with the class
instance identifiers.}

\item{colsToInclude}{The column names of the columns with the attributes.}

\item{output}{The filename to write the attributes to.}

\item{delimiterString}{When exporting to \code{.rock} format (as opposed to
\code{.yml} or \code{.yaml}), the delimiter to use to delimit the YAML.}

\item{attributeContainer}{The name of the root-level YAML node to store the
attributes in (the ROCK standard specifies this should
be \code{ROCK_attributes}).}

\item{col_renaming}{A named character vector where the names are the
current column names and the values are the new names to be used when
writing the corresponding attributes.}

\item{sheet}{The index (or name) of the worksheet containing the attributes.}
}
\value{
The produced YAML as a character vector (invisibly if \code{output}
is \code{NULL}).
}
\description{
This function takes a data frame and then writes them to a file in the
correct format of ROCK attributes.
}
\details{
Note that if the \code{output} file has the extension \code{.yml} or \code{.yaml}, the
delimiters passed in \code{delimiterString} will be ignored.
}
