% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericToBase30.R
\name{base30and36conversion}
\alias{base30and36conversion}
\alias{numericToBase30}
\title{Conversion between base10 and base30 & base36}
\usage{
numericToBase30(x)
}
\arguments{
\item{x}{The vector to convert (numeric for the \code{numericTo} functions,
character for the \code{base30to} and \code{base36to} funtions).}
}
\value{
The converted vector (numeric for the \code{base30to} and \code{base36to} funtions,
character for the \code{numericTo} functions).
}
\description{
The conversion functions from base10 to base30 are
used by the \code{\link[=generate_id]{generate_id()}} functions; the base36
functions are just left here for convenience.
}
\details{
The symbols to represent the 'base 30' system are the 0-9
followed by the alphabet without vowels but including the y. This
vector is available as \code{base30}.
}
\examples{
numericToBase30(654321);
base30toNumeric(numericToBase30(654321));
}
