% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igv.R
\name{bedval2igv}
\alias{bedval2igv}
\title{Export BED values for IGV Viewing}
\usage{
bedval2igv(bed, out_file = NULL, track_name = "val", col = "purple")
}
\arguments{
\item{bed}{BED file with \code{chrom}, \code{start}, \code{end} and \code{value} columns.}

\item{out_file}{Path to write output to. \strong{Needs to have an \code{igv} suffix}.}

\item{track_name}{Name of track to appear in IGV (default: "val").}

\item{col}{Colour name for the points (default: blue).}
}
\value{
Invisible list with the modified BED file and the track header, where:
\itemize{
\item \code{chrom}: chromosome
\item \code{start}: start coordinate
\item \code{end}: end coordinate
\item \code{name}: just an ID for the row (\code{chrom:start-end})
\item \code{val}: value
\item \code{header}: track information
}
}
\description{
Given a BED-like file with chromosome, start, end and value columns,
export a properly formatted file for viewing the values as a scatter plot
in \href{http://software.broadinstitute.org/software/igv/}{IGV}. Note that
\code{start} needs to be equal to \code{end}, since we're looking at specific SNPs,
not segments.
}
\examples{
\dontrun{
bed <- system.file("extdata", "COLO829_chr21_baf.tsv", package = "pebbles")
bedval2igv(bed, out_file = "~/Desktop/tmp/baf1.igv", track_name = "colo829_baf", col = "red")
}

}
