% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carry_over_values.R
\name{carry_over_values}
\alias{carry_over_values}
\title{Taking a vector, carry value over ('persistence')}
\usage{
carry_over_values(x, noId = "no_id")
}
\arguments{
\item{x}{The vector}

\item{noId}{The value to add for the first empty elements}
}
\value{
The vector with the carries over elements
}
\description{
This function takes a value, and then replaces
empty elements (NA or zero-length character values)
with the last non-empty preceding element it
encountered.
}
\examples{
rock::carry_over_values(
  c(
    NA, NA, 3, NA, NA, 7, NA, NA
  )
);
}
