% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat0.R
\name{cat0}
\alias{cat0}
\title{Concatenate to screen without spaces}
\usage{
cat0(..., sep = "")
}
\arguments{
\item{...}{The character vector(s) to print; passed to \link{cat}.}

\item{sep}{The separator to pass to \link{cat}, of course, \code{""} by default.}
}
\value{
Nothing (invisible \code{NULL}, like \link{cat}).
}
\description{
The cat0 function is to cat what paste0 is to paste; it
simply makes concatenating many strings without a separator easier.
}
\examples{
cat0("The first variable is '", names(mtcars)[1], "'.");
}
