% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cat_names}
\alias{cat_names}
\title{Names of awesome cats}
\format{Atomic character vector with 3 elements}
\usage{
cat_names
}
\description{
A vector containing names of cats.
Some are cereal killers.
}
\keyword{datasets}
