% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_source.R
\name{change_source}
\alias{change_source}
\title{Do something with a source}
\usage{
change_source(
  input,
  codes,
  func,
  filter = TRUE,
  output = NULL,
  decisionLabel = NULL,
  justification = NULL,
  justificationFile = rock::opts$get("justificationFile"),
  preventOverwriting = rock::opts$get("preventOverwriting"),
  encoding = rock::opts$get("encoding"),
  silent = rock::opts$get("silent"),
  ...
)
}
\arguments{
\item{input}{A file with a source or an object with a source
produced by one of the \code{loading_sources} functions.}

\item{func}{The function to apply.}

\item{filter}{Optionally, a filter to apply to specify a subset of the
source(s) to process (see \code{\link[=get_source_filter]{get_source_filter()}}).}

\item{output}{If specified, the coded source will be written here.}

\item{justification}{The justification for this action.}

\item{justificationFile}{If specified, the justification is appended to
this file. If not, it is saved to the \code{\link[justifier:workspace]{justifier::workspace()}}. This can
then be saved or displayed at the end of the R Markdown file or R script
using \code{\link[justifier:save_workspace]{justifier::save_workspace()}}.}

\item{preventOverwriting}{Whether to prevent overwriting existing files
when writing the files to \code{output}.}

\item{encoding}{The encoding to use.}

\item{silent}{Whether to be chatty or quiet.}

\item{...}{Other arguments to pass to \code{fnc}.}
}
\value{
Invisibly, the recoded source(s) or source(s) object.
}
\description{
This function contains the general set of actions that are always used
when doing something to change a source (e.g. check the input, document the
justification, etc). Users should normally never call this function.
}
\examples{
### Get path to example source
examplePath <-
  system.file("extdata", package="rock");

### Get a path to one example file
exampleFile <-
  file.path(examplePath, "example-1.rock");

### Parse single example source
loadedExample <- rock::load_source(exampleFile);

}
