% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_get_item.R
\name{ci_get_item}
\alias{ci_get_item}
\title{Get an item in a specific language}
\usage{
ci_get_item(nrm_spec, item_id, language)
}
\arguments{
\item{nrm_spec}{The Narrative Response Model specification.}

\item{item_id}{The identifier of the requested item.}

\item{language}{The language of the stimuli.}
}
\value{
A character value with the item.
}
\description{
This function takes a Narrative Response Model specification as used in
NRM-based cognitive interviews, and composes an item based on the
specified template for that item, the specified stimuli, and the
requested language.
}
