% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{circos_data}
\alias{circos_data}
\title{Chromosome information for circos}
\format{List with three elements:
\itemize{
\item db: start and end angles for displaying chromosomes in a circos plot
\item seg_name: chromosome names (1-22, X, Y)
\item seg_num: number of chromosomes (24)
}}
\source{
\url{https://bioconductor.org/packages/release/bioc/html/OmicCircos.html}
}
\usage{
data(circos_data)
}
\description{
Chromosome information for circos
}
\keyword{datasets}
