% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{circos_prep}
\alias{circos_prep}
\title{Generate Perl Circos Plot Files}
\usage{
circos_prep(outdir = "circos", manta = NULL, cnv = NULL)
}
\arguments{
\item{outdir}{Directory to write the files to.}

\item{manta}{Path to Manta VCF file.}

\item{cnv}{Path to copy number call file.}
}
\value{
Generates the required files for a Perl circos plot with copy number
variant segments and structural variant links. Returns these files
invisibly.
}
\description{
Generates required files for a Perl circos plot containing
Manta structural variants and CNVkit/FACETS/PURPLE/TitanCNA copy number variants.
}
\examples{
\dontrun{
manta <- system.file("extdata", "HCC2218_manta.vcf", package = "pebbles")
cnv <- system.file("extdata", "HCC2218_cnvkit-call.cns", package = "pebbles")
outdir <- "~/Desktop/tmp/circos"

circos_prep(outdir = outdir, manta = manta, cnv = cnv)
}
}
