% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_teams.R
\name{clean_teams}
\alias{clean_teams}
\title{Cleaning a transcript from Microsoft Teams}
\usage{
clean_teams(
  file,
  output = NULL,
  keepTime = FALSE,
  flagConfidence = TRUE,
  confFlag = c(0.7),
  preventOverwriting = rock::opts$get(preventOverwriting),
  encoding = rock::opts$get(encoding),
  silent = rock::opts$get(silent)
)
}
\arguments{
\item{output}{If not \code{NULL},
this is the name (and path) of the file in which to save the processed source (if it
\emph{is} \code{NULL}, a new file will be created with \verb{_clean} appended to the file name).}

\item{keepTime}{If true, retains the timestamp for each section at the start of the line}

\item{flagConfidence}{If true puts *** at the start of each line where the
automated transcribing is uncertain. Threshold for this flag is set by \code{confFlag}}

\item{confFlag}{sets the threshold below which a line is marked as Teams having
poor confidence in the transcription}

\item{input}{For \code{clean_teams} a path to a file that contains
the source transcription.}
}
\description{
This function takes a caption file from Microsoft Teams/Streams and
makes it a more readable/editable transcript.
}
\details{
This can then be proofread for transcription errors, and once ready, further
processed.

The output can be set-up to retain the time-stamps related to the recording.
If keepTime is TRUE a single timestamp (rather than a time range) starts each line.
By default, there is also a flag indicating lines where the transcription
confidence is low — ie they need more attention when manually cleaning.
confFlag sets the level at which lines are flagged.
}
\examples{
exampleSource <-
"WEBVTT

 NOTE duration:\"01:14:13.5060000\"

 NOTE language:en-us

 NOTE Confidence: 0.864107251167297

 1fbc54e1-43d1-4f2a-b639-ea8466e6f279
 00:00:02.260 --> 00:00:06.412
 Go back on WhatsApp. Yeah. So
 my first question is just a

 NOTE Confidence: 0.864107251167297

 2c90cf04-4e4b-4857-8b40-273d37b1e878
 00:00:06.412 --> 00:00:09.872
 little bit. How did you come
 to teach research methods?"

 clean_teams("exampleSource")


}
