% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv2granges.R
\name{cnv2gr}
\alias{cnv2gr}
\title{Convert cnv object to GenomicRanges}
\usage{
cnv2gr(cnv)
}
\arguments{
\item{cnv}{Object of class \code{cnv}}
}
\value{
A \code{GenomicRanges} object with the following columns:
\itemize{
\item seqnames: chromosome
\item ranges: start-end coordinates
\item tot_cn: total copy number estimate
}
}
\description{
Converts an object of class \code{cnv} to \code{GenomicRanges}
}
\examples{
cnv_fname <- system.file("extdata", "HCC2218_purple.cnv.tsv", package = "pebbles")
cnv <- prep_purple_seg(cnv_fname)

cnv2gr(cnv)

}
