% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igv.R
\name{cnv2igv}
\alias{cnv2igv}
\title{Export CNV Segments for IGV Viewing}
\usage{
cnv2igv(
  cnv,
  out_file = NULL,
  track_name = "cnv_segs",
  col1 = "darkgreen",
  col2 = "red",
  autoscale = "on"
)
}
\arguments{
\item{cnv}{Either a BED file with \code{chrom}, \code{start}, \code{end} and \code{tot_cn} columns, or a \code{cnv} object.}

\item{out_file}{Path to write output to. \strong{Needs to have a \code{bedgraph} suffix}.}

\item{track_name}{Name of track to appear in IGV (default: "cnv_segs").}

\item{col1}{Colour name for amplifications (default: darkgreen).}

\item{col2}{Colour name for deletions (default: red).}

\item{autoscale}{With autoscaling on, IGV automatically adjusts the
plot Y scale to the data range currently in view.
As the user pans and moves, this scaling continually adjusts.
Available options: on, off (default: on).}
}
\value{
Invisible list with the modified \code{cnv} object and the track header, where:
\itemize{
\item \code{chrom}: chromosome
\item \code{start}: start coordinate
\item \code{end}: end coordinate
\item \code{tot_cn - 2}: total copy number estimate minus 2 (so that CN2 will be the baseline)
\item \code{header}: track information
}
}
\description{
Exports an object of class \code{cnv} to a
\href{http://genome.ucsc.edu/goldenPath/help/bedgraph.html}{bedgraph} file for viewing
in \href{http://software.broadinstitute.org/software/igv/}{IGV} as bars.
}
\examples{
\dontrun{
cn_fname <- system.file("extdata", "HCC2218_purple.cnv.tsv", package = "pebbles")
cnv <- prep_purple_seg(cn_fname)
cnv2igv(cnv, out_file = "~/Desktop/tmp/cnv_segs4igv.bedgraph")
}

}
