% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_cnvkit.R
\name{cnvkit_depth2igv}
\alias{cnvkit_depth2igv}
\title{Write CNVkit cnr to wig}
\usage{
cnvkit_depth2igv(cnr, out_file = NULL, track_name = "val", col = "lightblue")
}
\arguments{
\item{cnr}{Path to CNVkit's \code{cnvkit.cnr} file, with \code{chromosome},
\code{start}, \code{end}, \code{gene}, \code{log2}, \code{depth} and \code{weight} columns.}

\item{out_file}{Path to write output to. \strong{Needs to have a \code{wig} suffix}.}

\item{track_name}{Name of track to appear in IGV (default: "val").}

\item{col}{Colour name for the bars (default: lightblue).}
}
\value{
Invisible list with the modified cnr file and the track header.
}
\description{
Reads the \code{cnvkit.cnr} file output by CNVkit and outputs its
\code{depth} column into a wiggle file for easy viewing in IGV.
}
\examples{
\dontrun{
cnr <- system.file("extdata", "HCC2218-cnvkit_chr22_1K.cnr", package = "pebbles")
cnvkit_depth2igv(cnr, out_file = "HCC2218.cov.wig", track_name = "HCC2218_coverage", col = "orange")
}

}
