% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_by_classInstance_barchart.R
\name{code_by_classInstance_barchart}
\alias{code_by_classInstance_barchart}
\title{Create a bar chart showing code occurrences for each class instance}
\usage{
code_by_classInstance_barchart(
  x,
  wrapLabels = 80,
  classId = "tssid",
  codesRegex = ".*",
  classInstanceRegex = ".*",
  classInstanceLab = NULL,
  codeLab = NULL,
  freqLab = "Count",
  plotTitle = "Heatmap",
  fillScale = ggplot2::scale_fill_viridis_c(),
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{x}{The object with the parsed coded source(s) as resulting from a
call to \code{\link[=parse_source]{parse_source()}} or \code{\link[=parse_sources]{parse_sources()}}.}

\item{wrapLabels}{Whether to wrap the labels; if not \code{NULL}, the
number of character to wrap at.}

\item{classId}{The identifier of the class for which the class instances should
form the rows (the class identifier is the bit to the left of the \code{=} or \code{:} in
class instance codings).}

\item{codesRegex}{A regular expression specifying which codes identifiers
to select for the columns.}

\item{classInstanceRegex}{A regular expression specifying which class instance
identifiers to select for the rows.}

\item{classInstanceLab}{Labels to use for the rows (\code{NULL} to omit the label).}

\item{codeLab}{Labels to use for the codes (\code{NULL} to omit the label).}

\item{freqLab}{Labels to use for the cell colors (\code{NULL} to omit the label).}

\item{plotTitle}{The title to use for the plot}

\item{fillScale}{Convenient way to specify the fill scale (the colours)}

\item{theme}{Convenient way to specify the \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} theme.}
}
\value{
The heatmap as a ggplot2 plot.
}
\description{
This function creates a bar chart with class instances (or sources) in rows,
plotting the code occurrences for each code in bars.
}
\examples{
examplePath <-
  file.path(
    system.file(package="rock"), 'extdata'
  );

parsedSources <- rock::parse_sources(
  examplePath,
  regex = "example-[123].rock"
);

### If no TSSIDs were configures, you can use
### source filenames instead like this:
rock::code_by_classInstance_heatmap(
  parsedSources,
  classId = "originalSource"
);
}
