% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_to_yaml.R
\name{codebook_to_yaml}
\alias{codebook_to_yaml}
\title{Produce the YAML for a ROCK codebook}
\usage{
codebook_to_yaml(x)
}
\arguments{
\item{x}{The ROCK codebook}
}
\value{
The YAML as a character vector
}
\description{
Produce the YAML for a ROCK codebook
}
\examples{
data(exampleCodebook_1, package="rock");
cat(
  codebook_to_yaml(
    exampleCodebook_1
  )
);
}
