% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeIds_to_codePaths.R
\name{codeIds_to_codePaths}
\alias{codeIds_to_codePaths}
\title{Replace code identifiers with their full paths}
\usage{
codeIds_to_codePaths(
  x,
  stripRootsFromCodePaths = rock::opts$get("stripRootsFromCodePaths")
)
}
\arguments{
\item{x}{A \code{rock_parsedSource} or \code{rock_parsedSources} object as returned by
a call to \code{\link[=parse_source]{parse_source()}} or \code{\link[=parse_sources]{parse_sources()}}.}

\item{stripRootsFromCodePaths}{Whether to strip the roots first (i.e. the
type of code)}
}
\value{
An adapted \code{rock_parsedSource} or \code{rock_parsedSources} object.
}
\description{
This function replaces the column names in the \code{mergedSourceDf} data frame
in a \code{rock_parsedSource} or \code{rock_parsedSources} object with the full paths
to those code identifiers.
}
