% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_to_nodes.R
\name{codes_to_nodes}
\alias{codes_to_nodes}
\title{Convert specifications of (deductive) codes into a hierarchical coding structure}
\usage{
codes_to_nodes(codes, silent = TRUE)
}
\arguments{
\item{codes}{The list of YAML fragments that contain information about codes,
as extracted from the sources by \code{\link[=extract_yaml_fragments]{extract_yaml_fragments()}}.}

\item{silent}{Whether to provide (\code{FALSE}) or suppress (\code{TRUE}) more detailed progress updates.}
}
\value{
a \code{\link[data.tree:Node]{data.tree::Node()}} object.
}
\description{
This function is called by \code{\link[=parse_sources]{parse_sources()}} to process the specification(s)
of codes as used for deductive coding into one hierarchical coding structure
(or coding tree).
}
\examples{
Add example here!
}
