% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingSchemes_get_all.R
\name{codingSchemes_get_all}
\alias{codingSchemes_get_all}
\title{Convenience function to get a list of all available coding schemes}
\usage{
codingSchemes_get_all()
}
\value{
A list of all available coding schemes
}
\description{
Convenience function to get a list of all available coding schemes
}
\examples{
rock::codingSchemes_get_all();
}
