% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncing_df_compress.R,
%   R/syncing_vector_compress.R
\name{syncing_df_compress}
\alias{syncing_df_compress}
\alias{syncing_vector_compress}
\title{Compress a vector or data frame}
\usage{
syncing_df_compress(x, newLength, sep = " ")

syncing_vector_compress(x, newLength, sep = " ")
}
\arguments{
\item{x}{The vector or data frame}

\item{newLength}{The new length}

\item{sep}{The separator used to glue together elements}
}
\value{
The shortened vector
}
\description{
Compress a vector or data frame
}
\examples{
rock::syncing_vector_compress(1:10, 3);
}
