% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToNumeric.R
\name{convertToNumeric}
\alias{convertToNumeric}
\title{Conveniently convert vectors to numeric}
\usage{
convertToNumeric(vector, byFactorLabel = FALSE)
}
\arguments{
\item{vector}{The vector to convert.}

\item{byFactorLabel}{When converting factors, whether to do this
by their label value (\code{TRUE}) or their level value (\code{FALSE}).}
}
\value{
The converted vector.
}
\description{
Tries to 'smartly' convert factor and character vectors to numeric.
}
\examples{
rock::convertToNumeric(as.character(1:8));
}
