% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{css}
\alias{css}
\title{Create HTML fragment with CSS styling}
\usage{
css(
  template = "default",
  includeBootstrap = rock::opts$get("includeBootstrap")
)
}
\arguments{
\item{template}{The template to load; either the name of one
of the ROCK templates (currently, only 'default' is available), or
the path and filename of a CSS file.}

\item{includeBootstrap}{Whether to include the default bootstrap CSS.}
}
\value{
A character vector with the HTML fragment.
}
\description{
Create HTML fragment with CSS styling
}
