% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_attributes.R
\name{expand_attributes}
\alias{expand_attributes}
\title{Expand categorical attribute variables to a series of dichotomous variables}
\usage{
expand_attributes(
  data,
  attributes,
  prefix = "",
  glue = "__",
  suffix = "",
  falseValue = 0,
  trueValue = 1,
  valueFirst = TRUE,
  append = TRUE
)
}
\arguments{
\item{data}{The data frame, normally the \verb{$mergedSources} data frame that
exists in the object returned by a call to \code{\link[=parse_sources]{parse_sources()}}.}

\item{attributes}{The name of the attribute(s) to expand.}

\item{prefix, suffix}{The prefix and suffix to add to the variables names
that are returned.}

\item{glue}{The glue to paste the first part ad the second part of the
composite variable name together.}

\item{falseValue, trueValue}{The values to set for rows that, respectively,
do not match and do match an attribute value.}

\item{valueFirst}{Whether to insert the attribute value first, or the
attribute name, in the composite variable names.}

\item{append}{Whether to append the columns to the supplied data
frame or not.}
}
\value{
A data.frame
}
\description{
Expand categorical attribute variables to a series of dichotomous variables
}
