% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncing_df_expand.R, R/syncing_vector_expand.R
\name{syncing_df_expand}
\alias{syncing_df_expand}
\alias{syncing_vector_expand}
\title{Expand a vector}
\usage{
syncing_df_expand(x, newLength, fill = TRUE)

syncing_vector_expand(x, newLength, fill = TRUE)
}
\arguments{
\item{x}{The vector}

\item{newLength}{The new length}

\item{fill}{Whether to duplicate elements to completely fill the new vector}
}
\value{
The expanded vector
}
\description{
Expand a vector
}
\examples{
rock::syncing_vector_expand(letters[1:10], 15);
rock::syncing_vector_expand(letters[1:10], 15, fill=FALSE);
}
