% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_attributes_from_parsedSources.R
\name{export_attributes_from_parsedSources}
\alias{export_attributes_from_parsedSources}
\title{Export the attributes in a parsed source(s) object to YAML}
\usage{
export_attributes_from_parsedSources(
  x,
  file = "ROCK_attributes.rock",
  path = NULL,
  preventOverwriting = rock::opts$get("preventOverwriting")
)
}
\arguments{
\item{x}{A \code{rock_parsedSource} or \code{rock_parsedSources} object.}

\item{file}{The filename to write to; pass \code{NULL} to only return the YAML as a character vector}

\item{path}{If not \code{NULL}, this will be combined with the \code{file} argument}
}
\value{
The YAML as a character vector
}
\description{
Export the attributes in a parsed source(s) object to YAML
}
