% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mergedSourceDf_to_file.R
\name{export_mergedSourceDf_to_csv}
\alias{export_mergedSourceDf_to_csv}
\alias{export_mergedSourceDf_to_csv2}
\alias{export_mergedSourceDf_to_xlsx}
\alias{export_mergedSourceDf_to_sav}
\title{Export a merged source data frame}
\usage{
export_mergedSourceDf_to_csv(
  x,
  file,
  exportArgs = list(fileEncoding = rock::opts$get("encoding")),
  preventOverwriting = rock::opts$get("preventOverwriting"),
  silent = rock::opts$get("silent")
)

export_mergedSourceDf_to_csv2(
  x,
  file,
  exportArgs = list(fileEncoding = rock::opts$get("encoding")),
  preventOverwriting = rock::opts$get("preventOverwriting"),
  silent = rock::opts$get("silent")
)

export_mergedSourceDf_to_xlsx(
  x,
  file,
  exportArgs = NULL,
  preventOverwriting = rock::opts$get("preventOverwriting"),
  silent = rock::opts$get("silent")
)

export_mergedSourceDf_to_sav(
  x,
  file,
  exportArgs = NULL,
  preventOverwriting = rock::opts$get("preventOverwriting"),
  silent = rock::opts$get("silent")
)
}
\arguments{
\item{x}{The object with parsed sources.}

\item{file}{The file to export to.}

\item{exportArgs}{Optionally, arguments to pass to the function to use to
export.}

\item{preventOverwriting}{Whether to prevent overwriting if the file already
exists.}

\item{silent}{Whether to be silent or chatty.}
}
\value{
Silently, the object with parsed sources.
}
\description{
Export a merged source data frame
}
