% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_codings_by_coderId.R
\name{extract_codings_by_coderId}
\alias{extract_codings_by_coderId}
\title{Extract the codings by each coder using the coderId}
\usage{
extract_codings_by_coderId(
  input,
  recursive = TRUE,
  filenameRegex = ".*",
  postponeDeductiveTreeBuilding = TRUE,
  ignoreOddDelimiters = FALSE,
  encoding = rock::opts$get(encoding),
  silent = rock::opts$get(silent)
)
}
\arguments{
\item{input}{The directory with the sources.}

\item{recursive}{Whether to also process subdirectories.}

\item{filenameRegex}{Only files matching this regular expression will be processed.}

\item{postponeDeductiveTreeBuilding}{Whether to build deductive code trees, or only
store YAML fragments.}

\item{ignoreOddDelimiters}{Whether to throw an error when encountering an odd number of
YAML delimiters.}

\item{encoding}{The encoding of the files to read.}

\item{silent}{Whether to be chatty or silent.}
}
\value{
An object with the read sources.
}
\description{
Extract the codings by each coder using the coderId
}
