% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rock_notes_from_text.R,
%   R/extract_rock_uids_from_text.R
\name{extract_rock_notes_from_text}
\alias{extract_rock_notes_from_text}
\alias{extract_rock_uids_from_text}
\title{Extract ROCK elements a text}
\usage{
extract_rock_notes_from_text(x)

extract_rock_uids_from_text(x)
}
\arguments{
\item{text}{The text as a character vector.}
}
\value{
An object containing the text, both in its original version
and in a version omitting the extracted elements, as well as the
extracted elements.
}
\description{
Extract ROCK elements a text
}
\examples{
### Get path to example source
examplePath <-
  system.file("extdata", package="rock");

### Get a path to one example file
exampleFile <-
  file.path(examplePath, "example-5.rock");

### Parse single example source
loadedExample <-
  rock::load_source(exampleFile);

### Extract notes
notes <-
  rock::extract_rock_notes_from_text(loadedExample);

notes$notes[[3]];

### Extract UIDs
UIDs <-
  rock::extract_rock_uids_from_text(loadedExample);
}
