% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_yaml_fragments.R
\name{extract_yaml_fragments}
\alias{extract_yaml_fragments}
\title{Extract all yaml fragments from a file}
\usage{
extract_yaml_fragments(file, text, delimiterRegEx = "^---$",
  ignoreOddDelimiters = FALSE, silent = TRUE)
}
\arguments{
\item{file}{The path to a file to scan.}

\item{text}{A character vector to scan, where every element should
represent one line in the file.}

\item{delimiterRegEx}{The regular expression used to locate YAML
fragments}

\item{ignoreOddDelimiters}{Whether to throw an error (FALSE) or
delete the last delimiter (TRUE) if an odd number of delimiters is
encountered.}

\item{silent}{Whether to be silent (TRUE) or informative (FALSE).}
}
\value{
A list of character vectors.
}
\description{
This function simply extracts all YAML fragments from a file,
returning a list of character vectors.
}
\details{
This function is called by \code{\link[=extract_dct_specs]{extract_dct_specs()}}; it is normally not
necessary to call it directly.
}
\examples{
extract_yaml_fragments(text=c("---", "First YAML fragment", "---",
                              "Outside of YAML",
                              "---", "Second fragment", "---",
                              "Also outside of YAML"));

}
