% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractNotes.R
\name{extractNotes}
\alias{extractNotes}
\title{Extract the notes from a text}
\usage{
extractNotes(text)
}
\arguments{
\item{text}{The text as a character vector}
}
\value{
An object containing the text, both in its original version
and in a cleaned version, as well as the notes.
}
\description{
Extract the notes from a text
}
\examples{
### Get path to example source
examplePath <-
  system.file("extdata", package="rock");

### Get a path to one example file
exampleFile <-
  file.path(examplePath, "example-5.rock");

### Parse single example source
loadedExample <-
  rock::load_source(exampleFile);

### Extract notes
notes <-
  rock::extractNotes(loadedExample);

}
