% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_classInstanceIds.R
\name{get_all_classInstanceIds}
\alias{get_all_classInstanceIds}
\title{Get a named list with vectors with all class instance identifiers}
\usage{
get_all_classInstanceIds(x)
}
\arguments{
\item{x}{An object with one or more parsed ROCK sources}
}
\value{
A named list of character vectors
}
\description{
Get a named list with vectors with all class instance identifiers
}
\examples{
### Get path to example source
examplePath <-
  system.file("extdata", package="rock");

### Get a path to one example file
exampleFile <-
  file.path(examplePath, "example-1.rock");

### Parse single example source
parsedExample <- rock::parse_source(exampleFile);

### Get all class instance identifiers
rock::get_all_classInstanceIds(
  parsedExample
);
}
