% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codeNode.R
\name{get_codeNode}
\alias{get_codeNode}
\title{Get the node from a code tree for a code identifier}
\usage{
get_codeNode(x, codeId)
}
\arguments{
\item{x}{The parsed sources object}

\item{codeId}{The code identifier of the code}
}
\value{
A \code{\link[data.tree:Node]{data.tree::Node()}} object
}
\description{
Get the node from a code tree for a code identifier
}
