% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataframe_from_nested_list.R
\name{get_dataframe_from_nested_list}
\alias{get_dataframe_from_nested_list}
\title{Return all values from a nested list in a dataframe}
\usage{
get_dataframe_from_nested_list(x, nestingIn = "children")
}
\arguments{
\item{x}{The nested list}

\item{nestingIn}{The name containing the nested lists}
}
\value{
A dataframe
}
\description{
Return all values from a nested list in a dataframe
}
\examples{
nestedList <-
  list(
    id = "x",
    value = "value for x",
    children = list(
      list(
        id = "y",
        value = "value for y"
      ),
      list(
        id = "z",
        value = "value for z"
      )
    )
  );
str(nestedList);
get_dataframe_from_nested_list(nestedList);
}
