% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parentCodeIds.R
\name{get_parentCodeIds}
\alias{get_parentCodeIds}
\title{Get the code identifiers a code's ascendents}
\usage{
get_parentCodeIds(
  x,
  childCodeId,
  parentOnly = TRUE,
  returnNodes = FALSE,
  returnPaths = FALSE,
  includeChildCode = FALSE
)
}
\arguments{
\item{x}{The parsed sources object}

\item{childCodeId}{The code identifier of the parent code}

\item{parentOnly}{Whether to only return the 'immediate' / 'direct' parent,
or the full ascendancy (including grand-parents, grand-grand parents,
etc etc).}

\item{returnNodes}{For \code{get_childCodeIds()}, set this to \code{TRUE} to return
a list of nodes, not just the code identifiers.}

\item{returnPaths}{Whether to return the paths (from the root) or the names.}

\item{includeChildCode}{Whether to include the child code
identifier in the result}
}
\value{
A character vector with code identifiers (or a list of nodes)
}
\description{
Get the code identifiers of all parents, or all ascendents (i.e. including
grand-parents, grand-grand-parents, etc) of a code with a given identifier.
}
