% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_UIDs.R
\name{has_UIDs}
\alias{has_UIDs}
\title{Check whether a source has UIDs}
\usage{
has_UIDs(x)
}
\arguments{
\item{x}{The source as a character vector.}
}
\value{
A single logical value.
}
\description{
A convenience function to quickly check whether a source has UIDs - or at
least more than one.
}
\examples{
exampleText <- c(
  "Lorem ipsum dolor sit amet, consectetur",
  "adipiscing elit. Nunc non commodo ex,",
  "ac varius mi. Praesent feugiat nunc",
  "eget urna euismod lobortis. Sed",
  "hendrerit suscipit nisl, ac tempus",
  "magna porta et. Quisque libero massa,",
  "tempus vel tristique lacinia, tristique",
  "in nulla. Nam cursus enim dui, non",
  "ornare est tempor eu. Vivamus et massa",
  "consectetur, tristique magna eget,",
  "viverra elit."
);

withUIDs <-
  rock::prepend_ids_to_source(
    exampleText
  );

rock::has_UIDs(
  exampleText
);

rock::has_UIDs(
  withUIDs
);
}
