% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say Hello to My Little Friend(s)}
\usage{
hello(x)
}
\arguments{
\item{x}{An atomic vector}
}
\value{
A character vector of length equal
to \code{x}, saying 'Hello' to each of its elements.
}
\description{
Says hello to all elements in \code{x}
}
\examples{
hello('UMCCR')
hello(c('Mollie', 'Walter', 'Dexter'))
}
