% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_source.R, R/load_sources.R
\name{load_source}
\alias{load_source}
\alias{load_sources}
\title{Load a source from a file or a string}
\usage{
load_source(input, encoding = "UTF-8", silent = FALSE)

load_sources(input, encoding = "UTF-8", silent = FALSE)
}
\arguments{
\item{input}{The filename or contents of the source
for \code{load_source} and the directory containing the
sources for \code{load_sources}.}

\item{encoding}{The encoding of the file(s).}

\item{silent}{Whether to be chatty or quiet.}
}
\value{
Invisibly, an R character vector of
classes \code{rock_source} and \code{character}.
}
\description{
These functions load one or more source(s) from a file or
a string and store it in memory for further processing.
Note that you'll probably want to clean the sources
first, using one of the \code{\link[=clean_sources]{clean_sources()}} functions,
and you'll probably want to add utterance identifiers
to each utterance using one of the \code{prepending_uids}
functions.
}
