% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_framework_table.R
\name{make_framework_table}
\alias{make_framework_table}
\title{Make a framework table}
\usage{
make_framework_table(
  parsedSources,
  classId,
  codeAggregation = "maxLeafs",
  codeAggregation_maxLeafs = 5,
  codeAggregation_maxDepth = 2,
  codeSelectionRegex = ".*",
  classInstanceSelectionRegex = ".*"
)
}
\arguments{
\item{parsedSources}{The object with parsed ROCK sources}

\item{classId}{The identifier of the class the instances of which should
form the framework table rows}

\item{codeAggregation}{How to aggregate codes (to prevent too many columns);
either \code{none} (do not aggregate codes), \code{maxLeafs} (the maximum number of
leaves (codes without children or 'sub-codes') to allow; with a higher
number, all codes are collapsed into their parent code) or \code{maxDepth}
(the maximum depth of codes to include; all codes at a higher 'depth'
(farther away from the root; the highest-level codes are level 1, their
children or 'sub-codes' are level 2, \emph{their} children or 'sub-codes' are
level 3, and so on) are collapsed into their parent code).}

\item{codeSelectionRegex}{A regular expression against which to match the codes to form
the framework table columns - this is matched against the original applied
codes, and so in combination with the code aggregation}

\item{classInstanceSelectionRegex}{A regular expression against which to match the
class instance identifiers to be included in the rows}

\item{maxLeafs}{When aggregating codes, the maximum number of leaves (codes
without children or 'sub-codes') to allow; with a higher number, all codes
are collapsed into their parent code}

\item{maxDepth}{When aggregating codes, (the maximum depth of codes to
include; all codes at a higher 'depth' (farther away from the root; the
highest-level codes are level 1, their children or 'sub-codes' are level 2,
\emph{their} children or 'sub-codes' are level 3, and so on) are collapsed into
their parent code.}
}
\value{
A framework table object; a list containing the framework
table as a data frame in its $framework_table slot
}
\description{
Make a framework table
}
