% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ROCK_codebook.R
\name{make_ROCK_codebook}
\alias{make_ROCK_codebook}
\title{Make a ROCK codebook}
\usage{
make_ROCK_codebook(
  metadata = NULL,
  codes = NULL,
  examples = NULL,
  aesthetics = NULL,
  relationships = NULL
)
}
\arguments{
\item{metadata}{The metadata, such as title, authors, and positionaly
statements, as a named list, where the names have to be the names as
defined in the ROCK standard.}

\item{codes}{The codes; either a character vector with code identifiers or a
data frame}

\item{examples, aesthetics, relationships}{Data frames with examples,
aesthetics, and relationships, as defined in the ROCK standard.}
}
\description{
With this function, you can create an empty ROCK codebook, populated
with the code identifiers you pass as \code{codes}. You can also pass data frames,
in the format of the spreadsheet ROCK codebook standard, to produce a ROCK
codebook object.
}
