% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_and_match_lists.R
\name{mix_and_match_lists}
\alias{mix_and_match_lists}
\title{Mix and match lists}
\usage{
mix_and_match_lists(x, y)
}
\arguments{
\item{x, y}{Named lists.}
}
\value{
The merged list.
}
\description{
Add elements from \code{y} to \code{x} only if they do not yet
exist in \code{x}. Without names, just \code{x} is returned.
}
\examples{
rock::mix_and_match_lists(
  list(a=1, b=4, c=7),
  list(b=9, d=9)
);
}
