% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr_in_interval.R
\name{nr_in_interval}
\alias{nr_in_interval}
\alias{nr_in_any_interval}
\alias{find_blocks_of_lines}
\title{Helper functions for note extraction}
\usage{
nr_in_interval(x, interval)

nr_in_any_interval(x, intervals)

find_blocks_of_lines(x)
}
\arguments{
\item{x}{A numeric vector of value 1 (for \code{\link[=nr_in_interval]{nr_in_interval()}} and more for
the other functions.}

\item{interval, intervals}{One or more intervals: sets of two or more line
numbers, with the minimum and maximum line numbers forming an internal}
}
\value{
A logical value or vector for \code{\link[=nr_in_interval]{nr_in_interval()}},
\code{\link[=nr_in_any_interval]{nr_in_any_interval()}}, and \code{\link[=nrs_in_any_interval]{nrs_in_any_interval()}}; a list for
\code{\link[=find_blocks_of_lines]{find_blocks_of_lines()}}
}
\description{
These are normally not used directly by end users; they're used
internally to extract notes.
}
\examples{
rock::nr_in_interval(5, 7:10);
rock::nr_in_interval(8, 7:10);

rock::nr_in_intervals(
  5,
  c(c(1,4), 7:10)
);

rock::find_blocks_of_lines(
  c(
    1, 2, 5, 7, 8, 9, 10,
    18, 20, 23, 24, 25, 30
  )
);
}
