% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_as_xl_date.R
\name{number_as_xl_date}
\alias{number_as_xl_date}
\title{Convert a number to a date using Excel's system}
\usage{
number_as_xl_date(x)
}
\arguments{
\item{x}{The number(s)}
}
\value{
The date(s)
}
\description{
Convert a number to a date using Excel's system
}
\examples{
preregr::number_as_xl_date(44113);
}
