% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padString.R
\name{padString}
\alias{padString}
\title{Padd a character vector}
\usage{
padString(x, width, padding = " ")
}
\arguments{
\item{x}{The character vector}

\item{width}{The width to pad to}

\item{padding}{The character to pad with}
}
\value{
\code{x} with padding appended to each element
}
\description{
Padd a character vector
}
\examples{
padString(
  c("One",
    "Two",
    "Three"
  ),
  width = 7
);
}
