% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sources.R
\name{parse_sources}
\alias{parse_sources}
\alias{print.rockParsedSources}
\alias{plot.rockParsedSources}
\title{Extract DCT specifications from all files in a directory}
\usage{
parse_sources(path, extension = "rock|dct", regex, codeRegexes = c(code
  = "\\\\[\\\\[([a-zA-Z0-9._>-]+)\\\\]\\\\]"), idRegexes = c(caseId =
  "\\\\[\\\\[cid=([a-zA-Z0-9._-]+)\\\\]\\\\]", stanzaId =
  "\\\\[\\\\[sid=([a-zA-Z0-9._-]+)\\\\]\\\\]"),
  autoGenerateIds = c("stanzaId"), sectionRegexes = c(paragraphs =
  "---paragraph-break---", secondary = "---<[a-zA-Z0-9]?>---"),
  inductiveCodingHierarchyMarker = ">", delimiterRegEx = "^---$",
  metadataContainers = c("metadata"), codesContainers = c("codes",
  "dct"), ignoreRegex = "^#", ignoreOddDelimiters = FALSE,
  encoding = "UTF-8", silent = TRUE)

\method{print}{rockParsedSources}(x, prefix = "### ", ...)

\method{plot}{rockParsedSources}(x, ...)
}
\arguments{
\item{path}{The path from where to read the DCT files.}

\item{extension}{The extension of the file; convenient alternative to
specifying a `regex`.}

\item{regex}{A regular expression; when provided, overrides the `extension`
argument to guide file selection in the `path`.}

\item{delimiterRegEx}{The regular expression specifying how the YAML fragments
specifying the constructs are delimited. Should normally never be changed.}

\item{metadataContainers, codesContainers}{A character vector listing the containers
of the metadata and code specifications in the YAML fragments. Because only
specifications are read that are stored in these containers, the files can contain
YAML fragments with other data, too, without interfering with the parsing of the
ROCK specifications.}

\item{ignoreOddDelimiters}{Whether to ignore a final odd delimiter, if
encountered.}

\item{silent}{Whether to provide information on progress.}
}
\value{
An object of class `dctRawSpecListSet` for processing by [parse_dct_specs()].
}
\description{
This function processes all DCT specifications in a directory.
}
\details{
This function is called by [process_dir()]; it is normally not
necessary to call this function directly.
}
\examples{
\dontrun{extract_dct_dir("A:/path/to/some/directory");
}
}
