% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsed_sources_to_ena_network.R
\name{parsed_sources_to_ena_network}
\alias{parsed_sources_to_ena_network}
\title{Title}
\usage{
parsed_sources_to_ena_network(x, unitCols, conversationCols,
  codesCols = x$convenience$codings,
  metadataCols = x$convenience$metadataVars)
}
\arguments{
\item{metadataCols}{}
}
\description{
Title
}
