% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2png.R
\name{pdf2png}
\alias{pdf2png}
\title{Transform PDF to PNG format}
\usage{
pdf2png(pdf, png = sub("pdf$", "png", pdf))
}
\arguments{
\item{pdf}{Path to input PDF file.}

\item{png}{Path to output PNG file.}
}
\value{
Path to output PNG file as a side effect.
}
\description{
Transforms PDF files to PNG files.
}
\details{
You need to make sure each file's suffix is
correct i.e. 'pdf' for the input PDF file, and 'png' for the
output PNG file. If you don't specify the 'png' argument,
by default it will output to the same path and with the same name as 'pdf',
but with a 'png' suffix.
}
\examples{
\dontrun{
pdf2png("img1.pdf", "img1.png")
pdf2png("img1.pdf")
}
}
