% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{plot_circos}
\alias{plot_circos}
\title{Generate Circos Plot}
\usage{
plot_circos(sv = NULL, cnv = NULL)
}
\arguments{
\item{sv}{An \code{sv} object}

\item{cnv}{A \code{cnv} object}
}
\value{
A circos plot with copy number variant segments
and structural variant links.
}
\description{
Generates a circos plot containing structural and copy number variants.
}
\details{
See the example for how to generate \code{sv} and \code{cnv} objects.
}
\examples{
\dontrun{
manta_vcf <- system.file("extdata", "HCC2218_manta.vcf", package = "pebbles")
cnvkit_seg <- system.file("extdata", "HCC2218_cnvkit-call.cns", package = "pebbles")

sv <- prep_manta_vcf(manta_vcf)
cnv <- prep_cnvkit_seg(cnvkit_seg)

pdf("/path/to/circos1.pdf", width = 7, height = 7)
plot_circos(sv = sv, cnv = cnv)
dev.off()
}
}
