% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{plot_circos_baf}
\alias{plot_circos_baf}
\title{Generate Perl Circos PURPLE BAF Plot}
\usage{
plot_circos_baf(purple_dir = NULL, out_dir = "circos_baf",
  tumor_name = NULL)
}
\arguments{
\item{purple_dir}{Directory where the circos files are located ('purple/circos').}

\item{out_dir}{Directory where the new circos plot will be written,
along with the input/config files.}

\item{tumor_name}{Tumor sample name.}
}
\value{
Generates a Perl circos plot in the outdir.
}
\description{
Generates a Perl circos plot containing
structural variant links, CNV calls, and BAF values from PURPLE.
}
\examples{
\dontrun{
purple_dir <- '/path/to/purple/circos'
out_dir <- '~/Desktop/tmp/circos_new'
tumor_name <- 'sampleA_tumor'

plot_circos_baf(purple_dir, out_dir, tumor_name)
}
}
