% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ideogram.R
\name{plot_ideogram}
\alias{plot_ideogram}
\title{Generate Ideogram Plot}
\usage{
plot_ideogram(chrom)
}
\arguments{
\item{chrom}{Chromosome character from 1-22, X or Y.}
}
\value{
An ideogram plot for the specified chromosome.
}
\description{
Generates an ideogram plot for the specified chromosome
}
\examples{
\dontrun{
plot_ideogram(chrom = "13")
}
}
