% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_piano.R
\name{plot_piano}
\alias{plot_piano}
\title{Generate Piano Plot}
\usage{
plot_piano(cnv_list)
}
\arguments{
\item{cnv_list}{A named list containing one or more \code{cnv} objects.
The element names will be used for plotting the facets so they need to be unique.
See the examples for usage.}
}
\value{
A piano plot with copy number variant segments across
multiple facets per sample or caller.
}
\description{
Generates a piano plot containing copy number variants from multiple samples
or callers.
}
\examples{
\dontrun{
cnvkit <- system.file("extdata", "HCC2218_cnvkit-call.cns", package = "pebbles")
facets <- system.file("extdata", "HCC2218_facets_cncf.tsv", package = "pebbles")
titan <- system.file("extdata", "HCC2218_titan.segs.tsv", package = "pebbles")
purple <- system.file("extdata", "HCC2218_purple.cnv.tsv", package = "pebbles")

cn_facets <- prep_facets_seg(facets)
cn_cnvkit <- prep_cnvkit_seg(cnvkit)
cn_titan <- prep_titan_seg(titan)
cn_purple <- prep_purple_seg(purple)
cnv_list <- list(cnvkit = cn_cnvkit, facets = cn_facets, purple = cn_purple, titan = cn_titan)

pdf("/path/to/piano1.pdf", width = 7, height = 7)
plot_piano(cnv_list)
dev.off()
}
}
