% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_general.R
\name{prep_cnv_circos}
\alias{prep_cnv_circos}
\title{Prepare CNV Segments for Perl Circos}
\usage{
prep_cnv_circos(cnv)
}
\arguments{
\item{cnv}{Path to CNV call file.}
}
\value{
A dataframe (\code{tibble}) with the following columns:
\itemize{
\item chrom: homo sapiens chromosome
\item start: start coordinate
\item end: end coordinate
\item value: total copy number estimate, minus 2
}
}
\description{
Exports CNV segment coordinates for plotting in Perl Circos.
}
\examples{
cn <- system.file("extdata", "HCC2218_cnvkit-call.cns", package = "pebbles")
prep_cnv_circos(cn)

}
