% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_cnvkit.R
\name{prep_cnvkit_circos}
\alias{prep_cnvkit_circos}
\title{Prepare CNVkit Segments for Perl Circos}
\usage{
prep_cnvkit_circos(cnvkit)
}
\arguments{
\item{cnvkit}{Path to CNVkit \code{sample-call.cns} file.}
}
\value{
A dataframe (\code{tibble}) with the following columns:
\itemize{
\item chrom: homo sapiens chromosome
\item start: start coordinate
\item end: end coordinate
\item value: total copy number estimate, minus 2
}
}
\description{
Reads the \code{sample-call.cns} file output by CNVkit and exports the
CNV segment coordinates for plotting in Perl Circos.
}
\examples{
cn <- system.file("extdata", "HCC2218_cnvkit-call.cns", package = "pebbles")
prep_cnvkit_circos(cn)

}
