% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_cnvkit.R
\name{prep_cnvkit_seg}
\alias{prep_cnvkit_seg}
\title{Read CNVkit CNV Segments}
\usage{
prep_cnvkit_seg(cnvkit)
}
\arguments{
\item{cnvkit}{Path to CNVkit \code{sample-call.cns} file.}
}
\value{
A dataframe (\code{tibble}) with the following columns:
\itemize{
\item chrom: chromosome
\item start: start coordinate
\item end: end coordinate
\item tot_cn: total copy number estimate
}
}
\description{
Reads the \code{sample-call.cns} file output by CNVkit and exports the
CNV segment coordinates.
}
\examples{
\dontrun{
prep_cnvkit_seg("/path/to/sample.cnvkit_call.cns")
}
}
