% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{prep_manta_vcf}
\alias{prep_manta_vcf}
\title{Prepare Manta VCF for Circos}
\usage{
prep_manta_vcf(vcf, filter_pass = TRUE)
}
\arguments{
\item{vcf}{Path to a bgzipped VCF file}

\item{filter_pass}{Keep only variants annotated with a PASS FILTER? (default: TRUE)}
}
\value{
A data.frame (tibble) with the genomic coordinates
of structural variants detected with Manta.
}
\description{
Prepares a Manta VCF for display in a circos plot. Requires bcftools to be
installed in your PATH.
}
\examples{
\dontrun{
prep_manta_vcf("/path/to/sample.vcf.gz")
}
}
