% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_manta.R
\name{prep_manta_vcf2}
\alias{prep_manta_vcf2}
\title{Prepare Manta VCF for Perl Circos}
\usage{
prep_manta_vcf2(vcf, ...)
}
\arguments{
\item{vcf}{Path to Manta VCF file. Can be compressed or not.}

\item{...}{Additional arguments for \code{prep_manta_vcf}.}
}
\value{
A dataframe (\code{tibble}) with the following fields from the VCF:
\itemize{
\item chrom1: hs\code{CHROM}
\item pos1: \code{POS} | \code{INFO/BPI_START}
\item pos1b: \code{POS} | \code{INFO/BPI_START}
\item chrom2: \code{CHROM} (for mate2 if BND)
\item pos2: \code{INFO/END} | \code{INFO/BPI_END} (for mate1 if BND)
\item pos2b: \code{INFO/END} | \code{INFO/BPI_END} (for mate1 if BND)
\item col: link colour
}
}
\description{
Prepares a Manta VCF for display in a Perl circos plot.
}
\examples{
vcf <- system.file("extdata", "HCC2218_manta.vcf", package = "pebbles")
prep_manta_vcf2(vcf)
}
