% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_purple.R
\name{prep_purple_seg}
\alias{prep_purple_seg}
\title{Read PURPLE CNV Segments}
\usage{
prep_purple_seg(purple)
}
\arguments{
\item{purple}{Path to PURPLE \code{sample-purple.cnv} file.}
}
\value{
A dataframe (\code{tibble}) with the following columns:
\itemize{
\item chrom: chromosome
\item start: start coordinate
\item end: end coordinate
\item tot_cn: total copy number estimate
}
}
\description{
Reads the \code{sample-purple.cnv} file output by PURPLE and exports the
CNV segment coordinates.
}
\examples{
purple <- system.file("extdata", "HCC2218_purple.cnv.tsv", package = "pebbles")
prep_purple_seg(purple)

}
