% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_titan.R
\name{prep_titan_seg}
\alias{prep_titan_seg}
\title{Read TitanCNA CNV Segments}
\usage{
prep_titan_seg(titan)
}
\arguments{
\item{titan}{Path to TitanCNA \code{sample-titan.segs.tsv} file.}
}
\value{
A dataframe (\code{tibble}) with the following columns:
\itemize{
\item chrom: chromosome
\item start: start coordinate
\item end: end coordinate
\item tot_cn: total copy number estimate
}
}
\description{
Reads the \code{sample-titan.segs.tsv} file output by TitanCNA and exports the
CNV segment coordinates.
}
\examples{
cn <- system.file("extdata", "HCC2218_titan.segs.tsv", package = "pebbles")
prep_titan_seg(cn)

}
