% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_truth.R
\name{prep_truth_seg}
\alias{prep_truth_seg}
\title{Read HCC2218 CNV Truth Set}
\usage{
prep_truth_seg(truth)
}
\arguments{
\item{truth}{Path to truth set.}
}
\value{
A dataframe (\code{tibble}) with the following columns:
\itemize{
\item chrom: chromosome
\item start: start coordinate
\item end: end coordinate
\item tot_cn: total copy number estimate
}
}
\description{
Reads the \code{HCC2218_truthset_cnv_bcbio.tsv} file in the \code{pebbles} package and
exports the CNV segment coordinates.
}
\examples{
cn <- system.file("extdata", "HCC2218_truthset_cnv_bcbio.tsv", package = "pebbles")
prep_truth_seg(cn)

}
