% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rock_graphList.R
\name{print.rock_graphList}
\alias{print.rock_graphList}
\title{Plot the graphs in a list of graphs}
\usage{
\method{print}{rock_graphList}(x, ...)
}
\arguments{
\item{x}{The list of graphs}

\item{...}{Any other arguments are passed to \code{\link[DiagrammeR:render_graph]{DiagrammeR::render_graph()}}.}
}
\value{
x, invisibly
}
\description{
Plot the graphs in a list of graphs
}
