% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qna.R
\name{qna}
\alias{qna}
\title{Parse a source and show its QNA network}
\usage{
qna(x)
}
\arguments{
\item{x}{A loaded source (with \code{\link[=load_source]{load_source()}}), the path to a
loaded source, or an already parsed source (with \code{\link[=parse_source]{parse_source()}}).}
}
\value{
A list with all produced graph(s) DiagrammeR objects and as Dot code.
}
\description{
Applies \code{\link[=parse_source]{parse_source()}} to a source, extract the produced network,
and shows it.
}
\examples{
### Get path to example source
examplePath <-
  system.file("extdata", package="rock");

### Read a souce coded with the Qualitative Network Approach
qnaExample <-
  file.path(
    examplePath,
    "network-example-1.rock"
  );

### Show the network
rock::qna(qnaExample);

}
