% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualchemy_project.R
\name{Qualchemy_project}
\alias{Qualchemy_project}
\title{R6 Class representing a {qualchemy} project}
\description{
R6 Class representing a {qualchemy} project

R6 Class representing a {qualchemy} project
}
\details{
A {qualchemy} project mostly just contains other objects.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{The project's metadata in a list}

\item{\code{classes}}{The project's classes as a character vector with the
class identifier of each class (e.g., "participantId" for the class
with participants)}

\item{\code{classInstances}}{A list of character vectors of class instance
identifiers (e.g., "participant_1", "question_7"), where each character
vector's name is the corresponding class identifier (e.g. "participantId",
"questionId", "locationId")}

\item{\code{codebook}}{A qualchemy codebook object}

\item{\code{sources}}{A list of qualchemy source objects}

\item{\code{codings}}{A list of qualchemy coding objects (a coding is the
application of a code to a part of a source)}

\item{\code{attributes}}{A named list of qualchemy attribute objects, where
each list element's name is the corresponding class identifier (e.g.
"participantId", "questionId", "locationId")}

\item{\code{extra}}{A named list of extra information.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Qualchemy_project-new}{\code{Qualchemy_project$new()}}
\item \href{#method-Qualchemy_project-add_source}{\code{Qualchemy_project$add_source()}}
\item \href{#method-Qualchemy_project-clone}{\code{Qualchemy_project$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qualchemy_project-new"></a>}}
\if{latex}{\out{\hypertarget{method-Qualchemy_project-new}{}}}
\subsection{Method \code{new()}}{
Create a new group object. Most of this text comes directly
from the TSV manual page at
\url{https://www.limesurvey.org/manual/Tab_Separated_Value_survey_structure}, so
please see that page for more details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qualchemy_project$new(
  metadata = "",
  classes = "",
  classInstances = list(),
  codebook = list(),
  sources = list(),
  codings = list(),
  attributes = list(),
  extra = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{A list with the project's metadata.}

\item{\code{classes}}{A character vector with the project's class identifiers.}

\item{\code{...}}{Any additional options, stored as a named list in the
\code{otherOptions} property by assigning \code{as.list(...)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Group} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qualchemy_project-add_source"></a>}}
\if{latex}{\out{\hypertarget{method-Qualchemy_project-add_source}{}}}
\subsection{Method \code{add_source()}}{
Add a source to a project.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qualchemy_project$add_source(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The path to the file with the source contents, or the source
contents as a character vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly, the \code{source} object that was just added. Note
that you can further modify this, which will modify the source object
"in" the project as well. This allows you to pipe the source
creation on to, for example, clean it.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qualchemy_project-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Qualchemy_project-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qualchemy_project$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
