% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualchemy_source.R
\name{Qualchemy_source}
\alias{Qualchemy_source}
\title{R6 Class representing a {qualchemy} source}
\description{
R6 Class representing a {qualchemy} source

R6 Class representing a {qualchemy} source
}
\details{
A {qualchemy} source: usually a text file that can be (or is) coded.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{contents}}{The source's contents}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Qualchemy_source-new}{\code{Qualchemy_source$new()}}
\item \href{#method-Qualchemy_source-load}{\code{Qualchemy_source$load()}}
\item \href{#method-Qualchemy_source-parent_project}{\code{Qualchemy_source$parent_project()}}
\item \href{#method-Qualchemy_source-clone}{\code{Qualchemy_source$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qualchemy_source-new"></a>}}
\if{latex}{\out{\hypertarget{method-Qualchemy_source-new}{}}}
\subsection{Method \code{new()}}{
Create a new source.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qualchemy_source$new(project = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The source contents.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Qualchemy_source} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qualchemy_source-load"></a>}}
\if{latex}{\out{\hypertarget{method-Qualchemy_source-load}{}}}
\subsection{Method \code{load()}}{
Import source contents from a file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qualchemy_source$load(x)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Invisibly, the \code{source} object itself.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qualchemy_source-parent_project"></a>}}
\if{latex}{\out{\hypertarget{method-Qualchemy_source-parent_project}{}}}
\subsection{Method \code{parent_project()}}{
Return the project that contains this source
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qualchemy_source$parent_project()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Invisibly, the \code{project} containing this source.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qualchemy_source-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Qualchemy_source-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qualchemy_source$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
