% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_source.R
\name{recode_source}
\alias{recode_source}
\title{Change a code into another code}
\usage{
recode_source(
  input,
  codes,
  output = NULL,
  justification = NULL,
  justificationFile = rock::opts$get("justificationFile"),
  preventOverwriting = rock::opts$get("preventOverwriting"),
  encoding = rock::opts$get("encoding"),
  silent = rock::opts$get("silent")
)
}
\arguments{
\item{input}{A file with a source (for \code{recode_source}), or a directory
with sources (for \code{recode_sources}), or an object with one source or
multiple sources as produced by one of the \code{loading_sources} functions.}

\item{codes}{A named list of character vectors, where each character vector
lists the codes to replace, and the name of that character vector is the
code to replace them with.}

\item{silent}{Whether to be chatty or quiet.}
}
\value{
Invisibly, the recoded source(s) or source(s) object.
}
\description{
These functions change a code into another code, and make it easy to
justify that decision.
}
\examples{
### Get path to example source
examplePath <-
  system.file("extdata", package="rock");

### Get a path to one example file
exampleFile <-
  file.path(examplePath, "example-1.rock");

### Parse single example source
loadedExample <- rock::load_source(exampleFile);

}
