% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock-package.R
\docType{package}
\name{rock-package}
\alias{rock-package}
\title{rock: Rocking R at UMCCR}
\description{
The \code{rock} package provides functions for several bioinformatic
analyses:
}
\details{
\itemize{
\item Foo
\item Bar
\item Baz
}
}
\section{Foo functions}{

The foo functions do this, this and that.
}

