% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_import_qrm_spec.R
\name{ci_import_qrm_spec}
\alias{ci_import_qrm_spec}
\alias{print.rock_ci_qrm}
\title{Import a Qualitative Response Model specification}
\usage{
ci_import_qrm_spec(
  x,
  read_ss_args = list(exportGoogleSheet = TRUE),
  silent = rock::opts$get("silent")
)

\method{print}{rock_ci_qrm}(x, ...)
}
\arguments{
\item{x}{A path to a file or an URL to a Google Sheet, passed
to \code{\link[=read_spreadsheet]{read_spreadsheet()}}.}

\item{read_ss_args}{A named list with arguments to pass to
\code{\link[=read_spreadsheet]{read_spreadsheet()}}.}

\item{silent}{Whether to be silent or chatty.}
}
\value{
A \code{rock_ci_qrm} object.
}
\description{
Import a Qualitative Response Model specification
}
