% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock_example_1.R
\docType{data}
\name{rock_example_1}
\alias{rock_example_1}
\title{Example 1}
\format{An object of class \code{character} of length 104.}
\usage{
rock_example_1
}
\description{
Example 1
}
\keyword{datasets}
