% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_from_codePaths.R
\name{root_from_codePaths}
\alias{root_from_codePaths}
\title{Get the roots from a vector with code paths}
\usage{
root_from_codePaths(x)
}
\arguments{
\item{x}{A vector of code paths.}
}
\value{
A vector with the root of each element.
}
\description{
Get the roots from a vector with code paths
}
\examples{
root_from_codePaths(
  c("codes>reason>parent_feels",
    "codes>reason>child_feels")
);
}
