% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr-package.R
\docType{package}
\name{rr-package}
\alias{rr-package}
\title{rr: Rocking R at UMCCR}
\description{
The \code{rr} package provides functions for several bioinformatic
analyses:
}
\details{
\itemize{
\item Foo
\item Bar
\item Baz
}
}
\section{Foo functions}{

The foo functions do this, this and that.
}

