% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplot.R
\name{samplot_commands}
\alias{samplot_commands}
\title{Prepare Samplot Commands from VCF}
\usage{
samplot_commands(vcf_file, bam_file, sample_nm, out_dir = NULL,
  other = "")
}
\arguments{
\item{vcf_file}{Path to the Manta VCF file (any SV VCF should work, really). Can be compressed or not.}

\item{bam_file}{Path(s) to BAM file(s) (atomic vector).}

\item{sample_nm}{Name(s) of BAM file(s) to display in plot (atomic vector).}

\item{out_dir}{Output directory.}

\item{other}{Other arguments to pass to Samplot (single string).}
}
\value{
A dataframe (\code{tibble}) with the following:
\itemize{
\item chrom1: \code{CHROM}
\item pos1: \code{POS} | \code{INFO/BPI_START}
\item chrom2: \code{CHROM} (for mate2 if BND)
\item pos2: \code{INFO/END} | \code{INFO/BPI_END} (for mate1 if BND)
\item svtype: \code{INFO/SVTYPE}
\item svlen: `pos2 - pos2`` if not BND, else 0
\item cl: Command line for Samplot
}
}
\description{
Prepares Samplot commands from a structural variant VCF and paths
to required BAM files.
}
\examples{

\dontrun{
vcf_file <- "path/to/vcf1.vcf.gz"
bam_file <- c("bam1", "bam2")
sample_nm <- c("s1", "s2")
samplot_commands(vcf_file, bam_file, sample_nm)
}

}
