% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_attribute_table.R
\name{show_attribute_table}
\alias{show_attribute_table}
\title{Show a table with all attributes in the RStudio viewer and/or console}
\usage{
show_attribute_table(
  x,
  output = rock::opts$get("tableOutput"),
  tableOutputCSS = rock::opts$get("tableOutputCSS")
)
}
\arguments{
\item{x}{A \code{rock_parsedSources} object (the result of a call to
\code{rock::parse_sources}).}

\item{output}{The output: a character vector with one or more
of "\code{console}" (the raw concatenated input, without conversion
to HTML), "\code{viewer}", which uses the RStudio viewer if available,
and one or more filenames in existing directories.}

\item{tableOutputCSS}{The CSS to use for the HTML table.}
}
\value{
\code{x}, invisibly, unless being knitted into R Markdown,
in which case a \code{\link[knitr:asis_output]{knitr::asis_output()}}-wrapped character vector is returned.
}
\description{
Show a table with all attributes in the RStudio viewer and/or console
}
