% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_long_lines.R
\name{split_long_lines}
\alias{split_long_lines}
\title{Split long lines}
\usage{
split_long_lines(
  x,
  length = 60,
  splitString = rock::opts$get("utteranceMarker")
)
}
\arguments{
\item{x}{The string (e.g. a source)}

\item{length}{The maximum length}

\item{splitString}{The character to use to split lines.}
}
\value{
A character vector.
}
\description{
This function splits long lines at a given number of characters,
keeping words intact. It's basically a wrapper around \code{\link[=strwrap]{strwrap()}}.
}
\examples{
cat(
  rock::split_long_lines(
    paste0(
      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. ",
      "Vestibulum et dictum urna. Donec neque nunc, lacinia vitae ",
      "varius vitae, pretium quis nibh. Aliquam pulvinar, lacus ",
      "sed varius vulputate, justo nibh blandit quam, ",
      "nec sollicitudin velit augue eget erat."
    )
  )
);
}
