% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripCodePathRoot.R
\name{stripCodePathRoot}
\alias{stripCodePathRoot}
\title{Strip the root from a code path}
\usage{
stripCodePathRoot(x)
}
\arguments{
\item{x}{A vector of code paths.}
}
\value{
The modified vector of code paths.
}
\description{
This function strips the root (just the first element) from a code path,
using the \code{codeTreeMarker} stored in the \link{opts} object as marker.
}
\examples{
stripCodePathRoot("codes>reason>parent_feels");
}
