% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_vector.R
\name{sync_vector}
\alias{sync_vector}
\title{Sync (expand or compress) a vector}
\usage{
sync_vector(x, newLength, sep = " ", fill = TRUE)
}
\arguments{
\item{x}{The vector}

\item{newLength}{The new length}

\item{sep}{The separator used to glue together elements}

\item{fill}{Whether to duplicate elements to completely fill the new vector}
}
\value{
The synced vector
}
\description{
Sync (expand or compress) a vector
}
\examples{
rock::sync_vector(letters[1:10], 15);
rock::sync_vector(letters[1:10], 5);
}
