% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapVector.R
\name{wrapVector}
\alias{wrapVector}
\title{Wrap all elements in a vector}
\usage{
wrapVector(x, width = 0.9 * getOption("width"), sep = "\\n", ...)
}
\arguments{
\item{x}{The character vector}

\item{width}{The number of}

\item{sep}{The glue with which to combine the new lines}

\item{...}{Other arguments are passed to \code{\link[=strwrap]{strwrap()}}.}
}
\value{
A character vector
}
\description{
Wrap all elements in a vector
}
\examples{
res <-
  rock::wrapVector(
    c(
      "This is a sentence ready for wrapping",
      "So is this one, although it's a bit longer"
    ),
    width = 10
  );

print(res);
cat(res, sep="\n");
}
