% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_attributes_to_dfs.R
\name{yaml_attributes_to_dfs}
\alias{yaml_attributes_to_dfs}
\title{Convert attributes from one or more YAML files to a list of data frames}
\usage{
yaml_attributes_to_dfs(x)
}
\arguments{
\item{x}{A character vector with one or more file paths}
}
\value{
A list of data frames
}
\description{
Convert attributes from one or more YAML files to a list of data frames
}
