% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_delimiter_indices.R
\name{yaml_delimiter_indices}
\alias{yaml_delimiter_indices}
\title{Get indices of YAML delimiters}
\usage{
yaml_delimiter_indices(x)
}
\arguments{
\item{x}{The character vector.}
}
\value{
A numeric vector.
}
\description{
Get indices of YAML delimiters
}
\examples{
yaml_delimiter_indices(
  c("not here",
    "---",
    "above this one",
    "but nothing here",
    "below this one, too",
    "---")
);
### [1] 2 6
}
