% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlify_cols_to_keyedvalues.R
\name{yamlify_cols_to_keyedvalues}
\alias{yamlify_cols_to_keyedvalues}
\title{Convert two columns in a data frame to a 'YAML' sequece of keyed values}
\usage{
yamlify_cols_to_keyedvalues(x, keyCol = 1, valueCol = 2, returnYAML = TRUE)
}
\arguments{
\item{x}{The data frame}

\item{keyCol}{The column holding the keys (names)}

\item{valueCol}{The column holding the values}

\item{returnYAML}{Whether to return 'YAML' (\code{TRUE}) or a list (\code{FALSE})}
}
\value{
'YAML' as character string or a list of lists
}
\description{
Convert two columns in a data frame to a 'YAML' sequece of keyed values
}
\examples{
cat(
  rock::yamlify_cols_to_keyedvalues(
    mtcars[1:2, ]
  )
);
}
