% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlify_rows_to_nodes.R
\name{yamlify_rows_to_nodes}
\alias{yamlify_rows_to_nodes}
\title{Convert rows in a data frame to 'YAML' nodes}
\usage{
yamlify_rows_to_nodes(
  x,
  keyCol = NULL,
  returnYAML = TRUE,
  colsToInclude = NULL,
  colsToOmit = NULL
)
}
\arguments{
\item{x}{The data frame}

\item{keyCol}{Optionally, a column used to name each YAML node. If set to
\code{NA}, the row names will be used as keys.}

\item{returnYAML}{Whether to return 'YAML' (\code{TRUE}) or a list (\code{FALSE})}
}
\value{
'YAML' as character string or a list of lists
}
\description{
Convert rows in a data frame to 'YAML' nodes
}
\examples{
cat(
  rock::yamlify_rows_to_nodes(
    mtcars[1:3, ]
  )
);
}
