% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_html_tags.R
\name{add_html_tags}
\alias{add_html_tags}
\title{Add HTML tags to a source}
\usage{
add_html_tags(x, codeRegexes = c(codes =
  "\\\\[\\\\[([a-zA-Z0-9._>-]+)\\\\]\\\\]"), idRegexes = c(caseId =
  "\\\\[\\\\[cid[=:]([a-zA-Z0-9._-]+)\\\\]\\\\]", stanzaId =
  "\\\\[\\\\[sid[=:]([a-zA-Z0-9._-]+)\\\\]\\\\]"),
  sectionRegexes = c(paragraphs = "---paragraph-break---", secondary =
  "---<[a-zA-Z0-9]?>---"),
  uidRegex = "\\\\[\\\\[uid[=:]([a-zA-Z0-9._-]+)\\\\]\\\\]",
  inductiveCodingHierarchyMarker = ">", codeClass = "code",
  idClass = "identifier", sectionClass = "sectionBreak",
  uidClass = "uid", utteranceClass = "utterance")
}
\arguments{
\item{x}{A character vector with the source}

\item{codeClass}{The classes to use for, respectively, codes,
identifiers (such as case identifiers or coder identifiers), section
breaks, utterance identifiers, and full utterances. All \code{<span>} elements
except for the full utterances, which are placed in \code{<div>} elements.}
}
\value{
The character vector witht he replacements made.
}
\description{
This function adds HTML tags to a source to allow pretty printing/viewing.
}
