% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_html_tags.R
\name{add_html_tags}
\alias{add_html_tags}
\title{Add HTML tags to a source}
\usage{
add_html_tags(
  x,
  codeClass = rock::opts$get(codeClass),
  idClass = rock::opts$get(idClass),
  sectionClass = rock::opts$get(sectionClass),
  uidClass = rock::opts$get(uidClass),
  utteranceClass = rock::opts$get(utteranceClass)
)
}
\arguments{
\item{x}{A character vector with the source}

\item{codeClass, idClass, sectionClass, uidClass, utteranceClass}{The classes
to use for, respectively, codes, identifiers (such as case identifiers or
coder identifiers), section breaks, utterance identifiers, and full
utterances. All \verb{<span>} elements except for the full utterances, which
are placed in \verb{<div>} elements.}
}
\value{
The character vector with the replacements made.
}
\description{
This function adds HTML tags to a source to allow pretty printing/viewing.
}
