% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.rock_source.R
\name{as.rock_source}
\alias{as.rock_source}
\title{Specify that something is a source}
\usage{
as.rock_source(x)
}
\arguments{
\item{x}{The source contents.}
}
\value{
A character vector with class \code{rock_source}.
}
\description{
This function converts an object to a character vector and marks it as
a ROCK source.
}
\examples{
exampleROCK <-
  rock::as.rock_source(c(
    "Some example text,",
    "and some more.       [[look_a_code]]",
    "And the end."));

### This can then be processed by other {rock}
### functions, for example:
rock::prettify_source(
  exampleROCK
);
}
