% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base30toNumeric.R, R/numericToBase30.R
\name{base30toNumeric}
\alias{base30toNumeric}
\alias{base30conversion}
\alias{numericToBase30}
\title{Conversion between base10 and base30}
\usage{
base30toNumeric(x)

numericToBase30(x)
}
\arguments{
\item{x}{The vector to convert (numeric for \code{numericToBase30},
character for \code{base30toNumeric}).}
}
\value{
The converted vector (numeric for \code{base30toNumeric}, character for \code{numericToBase30}).
}
\description{
The conversion functions from base10 to base30 and vice versa are
used by the \code{\link[=generate_uids]{generate_uids()}} functions.
}
\details{
The symbols to represent the 'base 30' system are the 0-9
followed by the alphabet without vowels but including the y. This
vector is available as \code{base30}.
}
\examples{
numericToBase30(654321);
base30toNumeric(numericToBase30(654321));
}
