% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{circos_prep}
\alias{circos_prep}
\title{Generate Perl Circos Plot Files}
\usage{
circos_prep(outdir = "circos", manta = NULL, cnv = NULL,
  genome = "hg19", ...)
}
\arguments{
\item{outdir}{Directory to write the files to.}

\item{manta}{Path to Manta VCF file.}

\item{cnv}{Path to copy number call file.}

\item{genome}{Genome assembly. "hg19" (default) or "hg38".
hg19 gets converted to GRCh37 automatically.}

\item{...}{Additional arguments for \code{prep_manta_vcf}.}
}
\value{
Generates the required files for a Perl circos plot with structural
variant links and/or copy number variant segments.
Returns the paths to these files invisibly.
}
\description{
Generates required files for a Perl circos plot containing
Manta structural variants and/or CNVkit/FACETS/PURPLE/TitanCNA
copy number variants.
}
\examples{
\dontrun{
manta <- system.file("extdata", "HCC2218_manta.vcf", package = "pebbles")
cnv <- system.file("extdata", "HCC2218_cnvkit-call.cns", package = "pebbles")
outdir <- "~/Desktop/tmp/circos"

circos_prep(outdir = outdir, manta = manta, cnv = cnv)
circos_prep(outdir = outdir, manta = manta) # no CNVs provided
circos_prep(outdir = outdir, manta = manta, genome = "hg19") # no CNVs provided
circos_prep(outdir = outdir, cnv = cnv) # no SVs provided
}
}
