% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{circos_prep}
\alias{circos_prep}
\title{Generate Perl Circos Plot Files}
\usage{
circos_prep(outdir = "circos", manta = NULL, cnvkit = NULL)
}
\arguments{
\item{outdir}{Directory to write the files to.}

\item{manta}{Path to Manta VCF file.}

\item{cnvkit}{Path to CNVkit \code{cnvkit-call.cns} file.}
}
\value{
Generates the required files for a Perl circos plot with copy number
variant segments and structural variant links.
}
\description{
Generates required files for a Perl circos plot containing
Manta structural variants and CNVkit copy number variants.
}
\examples{
\dontrun{
manta <- system.file("extdata", "HCC2218_manta.vcf", package = "pebbles")
cnvkit <- system.file("extdata", "HCC2218_cnvkit-call.cns", package = "pebbles")
outdir <- "circos"

circos_prep(outdir, manta, cnvkit)
}
}
