% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_source.R
\name{clean_source}
\alias{clean_source}
\title{Cleaning sources}
\usage{
clean_source(input, outputFile = NULL,
  replacements = list(c("[^\\\\.](\\\\.\\\\.)[^\\\\.]", "."),
  c("[^\\\\.](\\\\.\\\\.\\\\.\\\\.+)[^\\\\.]", "..."),
  c("\\\\n\\\\n\\\\n+", "\\n\\n")), extraReplacements = NULL,
  utteranceSplits = c("([[:alnum:]\\\\s]\\\\.(?!\\\\.\\\\.)\\\\s|[\\\\?\\\\!]+\\\\s|…\\\\s)"),
  utteranceMarker = "\\n", removeNewlines = FALSE,
  encoding = "UTF-8")
}
\arguments{
\item{input}{for \code{clean_source}, either a character vector containing the text
of the relevant source \emph{or} a path to a file that contains the source text;
for \code{clean_sources}, a path to a directory that contains the sources to clean.}

\item{outputFile}{If not \code{NULL}, this is the name (and path) of the file in
which to save the cleaned source.}

\item{replacements}{A list of two-element vectors, where the first element in each
vector contains a regular expression to search for in the source(s), and the second
element contains the replacement (these are passed as \code{perl} regular expressions; see
\code{\link{regex}} for more information). Instead of regular expressions, simple
words or phrases can also be entered of course (since those are valid regular
expressions).}

\item{extraReplacements}{To perform more replacements than the default set, these
can be conveniently specified in \code{extraReplacements}. This prevents you from having to
manually copypaste the list of defaults to retain it.}

\item{utteranceSplits}{This is a vector of regular expressions that specify where to
insert breaks between utterances in the source(s). Such breakes are specified using
\code{utteranceMarker}.}

\item{utteranceMarker}{How to specify breaks between utterances in the source(s). The
ROCK convention is to use a newline (\code{\n}).}

\item{removeNewlines}{Whether to remove all newline characters from the source before
starting to clean them.}

\item{encoding}{The encoding of the source(s).}
}
\value{
A character vector for \code{clean_source}, or a list of character vectors , for \code{clean_sources}.
}
\description{
Cleaning sources
}
