% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_freq_hist.R
\name{code_freq_hist}
\alias{code_freq_hist}
\title{Create a frequency histogram for codes}
\usage{
code_freq_hist(parsedSource, codes = ".*", sortByFreq = "decreasing")
}
\arguments{
\item{parsedSource}{A parsed source object.}

\item{codes}{A regular expression to select codes to include.}

\item{sortByFreq}{Whether to sort by frequency decreasingly
(\code{decreasing}, the default), increasingly (\code{increasing}),
or alphabetically (\code{NULL}).}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
Create a frequency histogram for codes
}
