% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_source.R, R/code_sources.R
\name{code_source}
\alias{code_source}
\alias{code_sources}
\title{Add one or more codes to one or more sources}
\usage{
code_source(input, codes, encoding = "UTF-8", silent = FALSE)

code_sources(input, codes, silent = FALSE)
}
\arguments{
\item{input}{The source, or list of sources, as
produced by one of the \code{loading_sources} functions.}

\item{codes}{A named character vector, where each element
is the code to be added to the matching utterance, and
the corresponding name is either an utterance identifier
(in which case the utterance with that identifier will be
coded with that code), a digit (in which case the utterance
at that line number in the source will be coded with that
code), or a regular expression, in which case all utterances
matching that regular expression will be coded with that
source.}

\item{silent}{Whether to be chatty or quiet.}
}
\value{
Invisibly, the coded source object.
}
\description{
These functions add codes to one or more
sources that were read with one of the
\code{loading_sources} functions.
}
