% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_coded_fragments.R
\name{collect_coded_fragments}
\alias{collect_coded_fragments}
\title{Create an overview of coded fragments}
\usage{
collect_coded_fragments(x, codes = ".*", context = 0,
  rawResult = FALSE, cleanUtterances = TRUE)
}
\arguments{
\item{x}{The parsed source(s) as provided by \code{rock::parse_source}
or \code{rock::parse_sources}.}

\item{codes}{The regular expression that matches the codes to include}

\item{context}{How many utterances before and after the target
utterances to include in the fragments.}

\item{rawResult}{Whether to returnt he raw result, a list of the
fragments, or one character value in markdown format.}

\item{cleanUtterances}{Whether to use the clean or the raw utterances
when constructing he fragments (the raw versions contain all codes).}
}
\value{
Either a list of character vectors, or a single character value.
}
\description{
Collect all coded utterances and optionally add some context
(utterances before and utterances after) to create ann overview
of all coded fragments per code.
}
\examples{
### Add later
}
