% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included_data_resources.R
\docType{data}
\name{exampleCodebook_1}
\alias{exampleCodebook_1}
\title{An very rudimentary example codebook specification}
\format{
An example of a codebook specification
}
\usage{
exampleCodebook_1
}
\description{
This is a simple and relatively short codebook
specification.
}
\keyword{datasets}
