% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_attributes_from_parsedSources.R
\name{export_attributes_from_parsedSources}
\alias{export_attributes_from_parsedSources}
\title{Export the attributes in a parsed source(s) object to YAML}
\usage{
export_attributes_from_parsedSources(
  x,
  file = "ROCK_attributes.rock",
  path = NULL,
  preventOverwriting = rock::opts$get("preventOverwriting")
)
}
\arguments{
\item{x}{A \code{rock_parsedSource} or \code{rock_parsedSources} object.}

\item{file}{The filename to write to; pass \code{NULL} to only return the YAML as a character vector}

\item{path}{If not \code{NULL}, this will be combined with the \code{file} argument}
}
\value{
The YAML as a character vector
}
\description{
Export the attributes in a parsed source(s) object to YAML
}
\examples{
### Get path to example source
examplePath <-
  system.file("extdata", package="rock");

### Get a path to one example file
exampleFile <-
  file.path(examplePath, "example-2.rock");

### Parse single example source
parsedExample <- rock::parse_source(exampleFile);

### Show attributes stord in YAML
### in the source
cat(
  rock::export_attributes_from_parsedSources(
    parsedExample,
    file = NULL
  )
);
}
