% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tssid.R
\name{generate_tssid}
\alias{generate_tssid}
\title{Generate a TSSID}
\usage{
generate_tssid(x = Sys.time())
}
\arguments{
\item{x}{The date and time to use.}
}
\value{
The tssid
}
\description{
A TSSID is a Timestamped Source Identifier. It is a practically
unique identifier for a source, conventionally the time and date
the source was produced (e.g. when the data were collected, for example
the time and date of an interview) in the UTC timezone and in ISO 8601
standard format.
}
\examples{
generate_tssid();
}
