% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_childCodeIds.R, R/get_descendentCodeIds.R
\name{get_childCodeIds}
\alias{get_childCodeIds}
\alias{get_descendentCodeIds}
\title{Get the code identifiers a code's descendents}
\usage{
get_childCodeIds(
  x,
  parentCodeId,
  returnNodes = FALSE,
  includeParentCode = FALSE
)

get_descendentCodeIds(x, parentCodeId, includeParentCode = FALSE)
}
\arguments{
\item{x}{The parsed sources object}

\item{parentCodeId}{The code identifier of the parent code}

\item{returnNodes}{For \code{get_childCodeIds()}, set this to \code{TRUE} to return
a list of nodes, not just the code identifiers.}

\item{includeParentCode}{Whether to include the parent code
identifier in the result}
}
\value{
A character vector with code identifiers (or a list of nodes)
}
\description{
Get the code identifiers of all children, or all descendents (i.e. including
grand-children, grand-grand-children, etc) of a code with a given identifier.
}
