% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_childCodeIds.R
\name{get_childCodeIds}
\alias{get_childCodeIds}
\title{Get the code identifiers of the children of a code with a given identifier}
\usage{
get_childCodeIds(x, parentCodeId, returnNodes = FALSE)
}
\arguments{
\item{x}{The parsed sources object}

\item{parentCodeId}{The code identifier of the parent code}

\item{returnNodes}{Set to \code{TRUE} to return a list of nodes, not just the
code identifiers}
}
\value{
A character vector with code identifiers (or a list of nodes)
}
\description{
Get the code identifiers of the children of a code with a given identifier
}
