% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_with_.R
\name{lines_with_codes}
\alias{lines_with_codes}
\alias{lines_with_breaks}
\alias{lines_with_anchors}
\alias{lines_with_uids}
\alias{lines_with_comments}
\alias{lines_with_nesting}
\alias{lines_with_whitespaceOnly}
\alias{remove_rock_from_text}
\alias{lines_with_data}
\title{Lines with specific contents}
\usage{
lines_with_codes(x)

lines_with_breaks(x)

lines_with_anchors(x)

lines_with_uids(x)

lines_with_comments(x)

lines_with_nesting(x)

lines_with_whitespaceOnly(x)

remove_rock_from_text(x)

lines_with_data(x)
}
\arguments{
\item{x}{The charactor vector to search.}
}
\value{
A logical vector.
}
\description{
These functions return a logical vector indicating on which lines the
specific content occurs.
}
\examples{
### Get path to example source
examplePath <-
  system.file("extdata", package="rock");

### Get a path to one example file
exampleFile <-
  file.path(examplePath, "example-5.rock");

### Parse single example source
loadedExample <-
  rock::load_source(exampleFile);

linesWithCodes <-
  lines_with_codes(loadedExample);
loadedExample[linesWithCodes];

linesWithBreaks <-
  lines_with_breaks(loadedExample);
loadedExample[linesWithBreaks];
}
