% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mata_determineTopics.R
\name{mata_determineTopics}
\alias{mata_determineTopics}
\title{Title}
\usage{
mata_determineTopics(
  x,
  classId_segmentation,
  topics = 5:10,
  parallel = FALSE,
  ciid_selection = NULL,
  max.em.its = 1000,
  heldout.seed = 101010,
  words_to_remove = NULL,
  min_word_nchar = 2,
  silent = rock::opts$get("silent")
)
}
\description{
Title
}
