% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\docType{data}
\name{opts}
\alias{opts}
\title{Options for the rock package}
\format{An object of class \code{list} of length 4.}
\usage{
rock::opts$set(...)
rock::opts$get(option, default=FALSE)
rock::opts$reset(...)
}
\arguments{
\item{...}{For \code{rock::opts$set}, the dots can be used to specify the options
to set, in the format \code{option = value}, for example, \code{utteranceMarker = "\n"}. For
\code{rock::opts$reset}, a list of options to be reset can be passed.}

\item{option}{The name of the option to set.}

\item{default}{The default value to return if the option has not been manually
specified.}
}
\description{
The \code{rock::opts} object contains three functions to set, get, and reset
options used by the rock package. Use \code{rock::opts$set} to set options,
\code{rock::opts$get} to get options, or \code{rock::opts$reset} to reset specific or
all options to their default values.
}
\keyword{datasets}
