% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{plot_circos}
\alias{plot_circos}
\title{Generate Perl Circos Plot}
\usage{
plot_circos(outdir = "circos", name = "x")
}
\arguments{
\item{outdir}{Directory where the files are located, and where the plot will
be written to.}

\item{name}{Prefix of plot file (suffix: \verb{_circos.png}) (default: \code{x}).}
}
\value{
Generates a Perl circos plot in the outdir.
}
\description{
Generates a Perl circos plot containing
Manta structural variants and/or CNV calls.
}
\examples{
\dontrun{
manta <- system.file("extdata", "HCC2218_manta.vcf", package = "pebbles")
cnv <- system.file("extdata", "HCC2218_cnvkit-call.cns", package = "pebbles")
outdir <- "circos"

circos_prep(outdir, manta, cnvkit)
plot_circos(outdir, name = "foo")
}
}
