% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{plot_circos}
\alias{plot_circos}
\title{Generate Circos Plot}
\usage{
plot_circos(sv, cnv)
}
\arguments{
\item{sv}{An \code{sv} object}

\item{cnv}{A \code{cnv} object}
}
\value{
A circos plot with copy number variant segments
and structural variant links
}
\description{
Generates a circos plot containing structural and copy number variants.
}
\details{
See the example for how to generate \code{sv} and \code{cnv} objects.
}
\examples{
\dontrun{
sv <- prep_manta_vcf("path/to/sample.manta.vcf.gz")
cnv <- prep_facets_seg("/path/to/sample.facets_emcncf.rds")

pdf("/path/to/circos1.pdf", width = 7, height = 7)
plot_circos(sv = sv, cnv = cnv)
dev.off()
}
}
