% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_piano.R
\name{plot_piano}
\alias{plot_piano}
\title{Generate Piano Plot}
\usage{
plot_piano(cnv_list, chromosomes = c(1:22, "X", "Y"), hide_x_lab = FALSE)
}
\arguments{
\item{cnv_list}{A named list containing one or more \code{cnv} objects.
The element names will be used for plotting the facets so they need to be unique.
See the examples for usage.}

\item{chromosomes}{Character vector of chromosomes to plot.}

\item{hide_x_lab}{Boolean. Should the X axis chromosome position labels be hidden or not?}
}
\value{
A piano plot with copy number variant segments across
multiple facets per sample or caller:
\itemize{
\item X axis: Chromosome Position
\item Y axis: Total Copy Number
\item Light green: amplifications (up to CN6)
\item Dark green: amplifications (greater than CN6, capped off at CN7)
\item Red: deletions
}
}
\description{
Generates a piano plot containing copy number variants from multiple samples
or callers.
}
\examples{
\dontrun{
cnvkit <- system.file("extdata", "HCC2218_cnvkit-call.cns", package = "pebbles")
facets <- system.file("extdata", "HCC2218_facets_cncf.tsv", package = "pebbles")
titan <- system.file("extdata", "HCC2218_titan.segs.tsv", package = "pebbles")
purple <- system.file("extdata", "HCC2218_purple.cnv.tsv", package = "pebbles")
truth <- system.file("extdata", "HCC2218_truthset_cnv_bcbio.tsv", package = "pebbles")

cn_facets <- prep_facets_seg(facets)
cn_cnvkit <- prep_cnvkit_seg(cnvkit)
cn_titan <- prep_titan_seg(titan)
cn_purple <- prep_purple_seg(purple)
cn_truth <- prep_truth_seg(truth)
cnv_list <- list(truth = cn_truth, cnvkit = cn_cnvkit, facets = cn_facets,
                 purple = cn_purple, titan = cn_titan)

plot_piano(cnv_list)
plot_piano(cnv_list, chromosomes = c(1:10))
}
}
