% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{prep_manta_vcf}
\alias{prep_manta_vcf}
\title{Prepare Manta VCF for Circos}
\usage{
prep_manta_vcf(vcf)
}
\arguments{
\item{vcf}{A vcfR object or a path to a VCF file}
}
\value{
A data.frame (tibble) with the genomic coordinates
of structural variants detected with Manta.
}
\description{
Prepares a Manta VCF for display in a circos plot
}
\examples{
\dontrun{
prep_manta_vcf(vcfr_object)
prep_manta_vcf("/path/to/vcf")
}
}
