% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_manta.R
\name{prep_manta_vcf}
\alias{prep_manta_vcf}
\title{Prepare Manta VCF for Circos}
\usage{
prep_manta_vcf(vcf, filter_pass = FALSE)
}
\arguments{
\item{vcf}{Path to Manta VCF file. Can be compressed or not.}

\item{filter_pass}{Keep only variants annotated with a PASS FILTER? (default: FALSE).}
}
\value{
A dataframe (\code{tibble}) with the following fields from the VCF:
\itemize{
\item chrom1: \code{CHROM}
\item pos1: \code{POS} | \code{INFO/BPI_START}
\item chrom2: \code{CHROM} (for mate2 if BND)
\item pos2: \code{INFO/END} | \code{INFO/BPI_END} (for mate1 if BND)
\item svtype: \code{INFO/SVTYPE}. Used for plotting.
}
}
\description{
Prepares a Manta VCF for display in a Circos plot.
}
\details{
This function uses bcftools (https://samtools.github.io/bcftools/bcftools.html)
or vcfR (https://github.com/knausb/vcfR) to read in the VCF file.
If you've got a large VCF file, or one with humongous gene annotations,
vcfR  will probably choke. A much quicker alternative is bcftools, but you need
to make sure it has been installed and that it can be found in the R session
via the PATH environmental variable. If you're an RStudio user, you can make
sure it recognises the user's PATH by opening the RStudio app via the terminal,
or perhaps following the suggestions here: https://stackoverflow.com/questions/31121645.
}
\examples{
vcf <- system.file("extdata", "HCC2218_manta.vcf", package = "pebbles")
prep_manta_vcf(vcf)

}
