% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_manta.R
\name{prep_manta_vcf}
\alias{prep_manta_vcf}
\title{Prepare Manta VCF for Circos}
\usage{
prep_manta_vcf(vcf, filter_pass = FALSE)
}
\arguments{
\item{vcf}{Path to bgzipped Manta VCF file. Needs to have \code{vcf.gz} suffix.}

\item{filter_pass}{Keep only variants annotated with a PASS FILTER? (default: FALSE).}
}
\value{
A dataframe (\code{tibble}) with the following fields from the VCF:
\itemize{
\item chrom1: \code{CHROM}
\item pos1: \code{POS} | \code{INFO/BPI_START}
\item chrom2: \code{CHROM} (for mate2 if BND)
\item pos2: \code{INFO/END} | \code{INFO/BPI_END} (for mate1 if BND)
\item svtype: \code{INFO/SVTYPE}. Used for plotting.
}
}
\description{
Prepares a Manta VCF for display in a circos plot.
}
\details{
VCF needs to be bgzipped.
Make sure bcftools has been installed and that it can be
found in the R session via the PATH environmental variable.
}
\examples{
vcf <- system.file("extdata", "HCC2218_manta.vcf", package = "pebbles")
prep_manta_vcf(vcf)

}
