% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepend_ids_to_source.R,
%   R/prepend_ids_to_sources.R
\name{prepend_ids_to_source}
\alias{prepend_ids_to_source}
\alias{prepend_ids_to_sources}
\title{Prepending unique utterance identifiers}
\usage{
prepend_ids_to_source(input, output = NULL, encoding = "UTF-8")

prepend_ids_to_sources(input, output = NULL, encoding = "UTF-8")
}
\arguments{
\item{input}{The filename or contents of the source.}

\item{output}{The filename where to write the resulting file.}

\item{encoding}{The encoding of the file.}
}
\value{
The source with prepended uids, either invisible (if \code{output}
if specified) or visibly (if not).
}
\description{
This function prepending unique utterance identifiers to each
utterance (line) in a source. Note that you'll probably want
to clean the sources using \code{\link[=clean_sources]{clean_sources()}} first.
}
