% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepend_ids_to_source.R,
%   R/prepend_ids_to_sources.R
\name{prepend_ids_to_source}
\alias{prepend_ids_to_source}
\alias{prepend_ids_to_sources}
\title{Prepending unique utterance identifiers}
\usage{
prepend_ids_to_source(input, output = NULL, origin = Sys.time(),
  preventOverwriting = TRUE, encoding = "UTF-8", silent = FALSE)

prepend_ids_to_sources(input, output = NULL, preventOverwriting = TRUE,
  encoding = "UTF-8", origin = Sys.time(), silent = FALSE)
}
\arguments{
\item{input}{The filename or contents of the source
for \code{prepend_ids_to_source} and the directory containing the
sources for \code{prepend_ids_to_sources}.}

\item{output}{The filename where to write the resulting file for
\code{prepend_ids_to_source} and the directory where to write the
resulting files for \code{prepend_ids_to_sources}}

\item{origin}{The time to use for the first identifier.}

\item{encoding}{The encoding of the file(s).}

\item{silent}{Whether to be chatty or quiet.}
}
\value{
The source with prepended uids, either invisible (if \code{output}
if specified) or visibly (if not).
}
\description{
This function prepending unique utterance identifiers to each
utterance (line) in a source. Note that you'll probably want
to clean the sources using \code{\link[=clean_sources]{clean_sources()}} first.
}
