% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_manta.R
\name{read_manta_vcf}
\alias{read_manta_vcf}
\title{Read Manta VCF}
\usage{
read_manta_vcf(vcf)
}
\arguments{
\item{vcf}{Path to bgzipped Manta VCF file. Needs to have \code{vcf.gz} suffix.}
}
\value{
A dataframe (\code{tibble}) with the following fields from the VCF:
\itemize{
\item chrom1: \code{CHROM}
\item pos1: \code{POS} | \code{INFO/BPI_START}
\item pos2: \code{INFO/END} | \code{INFO/BPI_END}
\item id: \code{ID}
\item mateid: \code{INFO/MATEID}
\item svtype: \code{INFO/SVTYPE}
\item filter: \code{FILTER}
}
}
\description{
Read main columns of interest from Manta VCF using bcftools
}
\details{
VCF needs to be bgzipped.
Make sure bcftools has been installed and that it can be
found in the R session via the PATH environmental variable.
}
\examples{
\dontrun{
read_manta_vcf("/path/to/sample.manta.vcf.gz")
}

}
