% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_manta.R
\name{read_manta_vcf}
\alias{read_manta_vcf}
\title{Read Manta VCF}
\usage{
read_manta_vcf(vcf)
}
\arguments{
\item{vcf}{Path to Manta VCF file (\code{.vcf.gz} or \code{.vcf}).}
}
\value{
A dataframe (\code{tibble}) with the following fields from the VCF:
\itemize{
\item chrom1: \code{CHROM} (remove \code{chr} prefix (if any) for hg38 compatibility)
\item pos1: \code{POS} | \code{INFO/BPI_START}
\item pos2: \code{INFO/END} | \code{INFO/BPI_END}
\item id: \code{ID}
\item mateid: \code{INFO/MATEID}
\item svtype: \code{INFO/SVTYPE}
\item filter: \code{FILTER}
}
}
\description{
Read main columns of interest from Manta VCF using bcftools or bedr
}
\details{
Uses bcftools (https://samtools.github.io/bcftools/bcftools.html)
or bedr (https://cran.r-project.org/web/packages/bedr/index.html) to read
in the VCF file.
}
\examples{
vcf <- system.file("extdata", "HCC2218_manta.vcf", package = "pebbles")
vcf2 <- system.file("extdata", "manta_no_bpi.vcf", package = "pebbles")
rock:::read_manta_vcf(vcf)
rock:::read_manta_vcf(vcf2)

}
