% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove__from_text.R
\name{remove_rock_codes_from_text}
\alias{remove_rock_codes_from_text}
\alias{remove_rock_breaks_from_text}
\alias{remove_rock_uids_from_text}
\alias{remove_rock_anchors_from_text}
\alias{remove_rock_comments_from_text}
\alias{remove_rock_nesting_from_text}
\alias{remove_rock_from_text}
\title{Removes ROCK elements from text}
\usage{
remove_rock_codes_from_text(x)

remove_rock_breaks_from_text(x)

remove_rock_uids_from_text(x)

remove_rock_anchors_from_text(x)

remove_rock_comments_from_text(x)

remove_rock_nesting_from_text(x)

remove_rock_from_text(x)
}
\arguments{
\item{x}{The text as a character vector.}
}
\value{
The character vector.
}
\description{
Removes ROCK elements from text
}
\examples{
### Get path to example source
examplePath <-
  system.file("extdata", package="rock");

### Get a path to one example file
exampleFile <-
  file.path(examplePath, "example-5.rock");

### Parse single example source
loadedExample <-
  rock::load_source(exampleFile);

### Extract notes
rock::remove_rock_codes_from_text(loadedExample);

}
