% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock-package.R
\docType{package}
\name{rock-package}
\alias{rock}
\alias{rock-package}
\title{rock: Reproducible Open Coding Kit}
\description{
The Reproducible Open Coding Kit ('ROCK', and this package, 'rock') was developed to facilitate reproducible and open coding, specifically geared towards qualitative research methods. Although it is a general-purpose toolkit, three specific applications have been implemented, specifically an interface to the 'rENA' package that implements Epistemic Network Analysis ('ENA'), means to process notes from Cognitive Interviews ('CIs'), and means to work with decentralized construct taxonomies ('DCTs'). The 'ROCK' and this 'rock' package are described in the ROCK book \url{https://rockbook.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rock.opens.science}
  \item Report bugs at \url{https://gitlab.com/r-packages/rock/-/issues}
}

}
\author{
\strong{Maintainer}: Gjalt-Jorn Ygram Peters \email{rock@opens.science} (\href{https://orcid.org/0000-0002-0336-9589}{ORCID})

Authors:
\itemize{
  \item Szilvia Zorgo (\href{https://orcid.org/0000-0002-6916-2097}{ORCID}) [contributor]
}

}
\keyword{internal}
