% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr-package.R
\docType{package}
\name{rr-package}
\alias{rr}
\alias{rr-package}
\title{rr: Rocking R for UMCCR}
\description{
Contains functions for bioinformatics analyses
}
\author{
\strong{Maintainer}: Peter Diakumis \email{peterdiakumis@gmail.com}

}
\keyword{internal}
