% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snoe_plot.R
\name{snoe_plot}
\alias{snoe_plot}
\title{Soft Non-numeric Occurrence Estimation (SNOE) plot}
\usage{
snoe_plot(
  x,
  codes = ".*",
  matchRegexAgainstPaths = TRUE,
  estimateWithin = NULL
)
}
\arguments{
\item{x}{A parsed source(s) object.}

\item{codes}{A regular expression to select codes to include, or,
alternatively, a character vector with literal code idenfitiers.}

\item{estimateWithin}{The column specifying within what to count.}

\item{sortByFreq}{Whether to sort by frequency decreasingly
(\code{decreasing}, the default), increasingly (\code{increasing}),
or alphabetically (\code{NULL}).}

\item{forceRootStripping}{Force the stripping of roots, even if they are
different.}

\item{ggplot2Theme}{Can be used to specify theme elements for the plot.}

\item{silent}{Whether to be chatty or silent.}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
Soft Non-numeric Occurrence Estimation (SNOE) plot
}
\examples{
### Get path to example source
examplePath <-
  system.file("extdata", package="rock");

### Get a path to one example file
exampleFile <-
  file.path(examplePath, "example-1.rock");

exampleFile <-
  readLines(
    "https://codeberg.org/explicate/ehps-2024-pluriformity-uniformity-roundtable/raw/branch/main/data-coded/sourceId-20240905T0800Z---coderId-consensus-1.rock"
  );

### Load example source
loadedExample <- rock::parse_source(exampleFile);

### Show code frequencies
rock::snoe_plot(
  loadedExample
);
}
