% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_streams.R
\name{sync_streams}
\alias{sync_streams}
\title{Synchronize multiple streams}
\usage{
sync_streams(
  x,
  primaryStream,
  anchorsCol = rock::opts$get("anchorsCol"),
  sourceId = rock::opts$get("sourceId"),
  streamId = rock::opts$get("streamId")
)
}
\arguments{
\item{x}{The object with the parsed sources.}

\item{primaryStream}{The identifier of the primary stream.}

\item{anchorsCol}{The column containing the anchors.}

\item{sourceId}{The column containing the source identifiers.}

\item{streamId}{The column containing the stream identifiers.}
}
\description{
This function maps the codes from multiple streams onto a primary stream.
}
