% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_regex_matches.R
\name{which_regex_matches}
\alias{which_regex_matches}
\title{Return which regular expression of several matches a character value}
\usage{
which_regex_matches(
  pattern,
  x,
  ignore.case = FALSE,
  perl = TRUE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\arguments{
\item{pattern}{The character vector with regular expressions to match against}

\item{x}{The text to match against the regular expressions}

\item{ignore.case}{Whether to be case sensitive (see \code{\link[base:grep]{base::grepl()}}).}

\item{perl}{Whether \code{pattern} specifies Perl regexes (see \code{\link[base:grep]{base::grepl()}}).}

\item{fixed}{Whether \code{pattern} is a regex or fixed string (see \code{\link[base:grep]{base::grepl()}}).}

\item{useBytes}{See \code{\link[base:grep]{base::grepl()}}.}
}
\value{
A logical vector or array (of \code{x} has more than one element)
}
\description{
Return which regular expression of several matches a character value
}
\examples{
pattern <-
  c("[0-9]",
    "[A-Z]",
    "[a-z]");

rock::which_regex_matches(
  pattern,
  "42"
);

rock::which_regex_matches(
  pattern,
  "forty-two"
);

rock::which_regex_matches(
  pattern,
  c(42, "forty-two")
);
}
