% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_regex_matches.R
\name{which_regex_matches}
\alias{which_regex_matches}
\title{Return which regular expression of several matches a character value}
\usage{
which_regex_matches(
  pattern,
  x,
  ignore.case = FALSE,
  perl = TRUE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\arguments{
\item{useBytes}{}
}
\description{
Return which regular expression of several matches a character value
}
\examples{
pattern <-
  c("[0-9]",
    "[A-Z]",
    "[a-z]");

rock::which_regex_matches(
  pattern,
  "42"
);

rock::which_regex_matches(
  pattern,
  "forty-two"
);

rock::which_regex_matches(
  pattern,
  c(42, "forty-two")
);
}
